### 2022-09-17 #60 @mobigroup
PyGMTSAR (Python GMTSAR)
Submitting Author: Alexey Pechnikov (informally) or Aleksei Pechnikov (by passport) (pechnikov@mobigroup.ru)
All current maintainers: (pechnikov@mobigroup.ru)
Package Name: PyGMTSAR
One-Line Description of Package: PyGMTSAR (Python GMTSAR) is an open source project and Python package that provides easy and fast Sentinel-1 Satellite Interferometry for everyone
Repository Link:  https://github.com/mobigroup/gmtsar
Version submitted:   2022.9.14.3 (https://pypi.org/project/pygmtsar/)
Version accepted: TBD

---

PyGMTSAR (Python GMTSAR) is an open source project and Python package that provides easy and fast Sentinel-1 Satellite Interferometry for everyone! While it's pure Python package under the hood it uses GMTSAR binary tools which should be installed.

	- [x ] Data retrieval
	- [x ] Data extraction
	- [x ] Reproducibility
	- [x ] Geospatial
	- [x ] Education

> *Please fill out a pre-submission inquiry before submitting a data visualization package. For more info, see [notes on categories][NotesOnCategories] of our guidebook.*

As I know the project used in research community, high school teachers and students (see GitHub followers for more details). And it's really available for everyone in just one click by web link, try the Live Google Colab examples on the GitHub project page.

The goal of the project is easy and fast satellite interferometry (InSAR) interactive and batch processing in Python scripts and Jupyter Notebooks for Sentinel-1 radar scenes everywhere as on localhost as on cloud environments like to Google Cloud VM and AI Notebooks and Amazon EC2 and even on free of charge cloud environment Google Colab.

- [x ] does not violate the Terms of Service of any service it interacts with.
- [x ] has an [OSI approved license][OsiApprovedLicense].
- [x ] contains a README with instructions for installing the development version.
- [x ] includes documentation with examples for all functions.
- [x ] contains a vignette with examples of its essential functions and uses.
- [x ] has a test suite.

- [x ] Do you wish to automatically submit to the [Journal of Open Source Software][JournalOfOpenSourceSoftware]? If so:

- [x ] The package has an **obvious research application** according to JOSS's definition in their [submission requirements][JossSubmissionRequirements]. Be aware that completing the pyOpenSci review process **does not** guarantee acceptance to JOSS. Be sure to read their submission requirements (linked above) if you are interested in submitting to JOSS.
- [x ] The package is not a "minor utility" as defined by JOSS's [submission requirements][JossSubmissionRequirements]: "Minor ‘utility’ packages, including ‘thin’ API clients, are not acceptable." pyOpenSci welcomes these packages under "Data Retrieval", but JOSS has slightly different criteria.

- [x ] I agree to abide by [pyOpenSci's Code of Conduct][PyOpenSciCodeOfConduct] during the review process and in maintaining my package should it be accepted.

[Editor and review templates can be found here][Templates]

[NotesOnCategories]: https://www.pyopensci.org/contributing-guide/open-source-software-peer-review/aims-and-scope.html?highlight=data#notes-on-categories

[Templates]: https://www.pyopensci.org/contributing-guide/appendices/templates.html
