### 2022-09-17 f910612
Remove repeated installation instruction
