### 2022-09-18 cd5dd9a
Enhance GitHub Pages layout for iPad mini
