### 2022-09-18 7e1f6d5
Add CI test for GMTSAR dataset S1A_Stack_CPGF_T173
