### 2022-09-18 a07281e
Simplify Python library dependencies list and minimize GitHub pages layout javascript code
