### 2022-09-18 7774ddb
Add comment how to download the test dataset
