### 2022-09-18 e7579a5
Add CI tests workflow
