+++ .github/workflows/ubuntu.yml
@@ -1,0 +1,86 @@
+# This workflow will install Python dependencies, run tests and lint with a single version of Python
+# For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
+
+name: Ubuntu tests
+
+on:
+  push:
+    branches: [ "master" ]
+  pull_request:
+    branches: [ "master" ]
+
+permissions:
+  contents: read
+
+jobs:
+  S1A_Stack_CPGF_T173:
+
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v3
+    - name: Set up Python 3.10
+      uses: actions/setup-python@v3
+      with:
+        python-version: "3.10"
+    - name: Install dependencies
+      run: |
+        uname -a
+        #GMTSAR=/usr/local/GMTSAR
+        #GIT=https://github.com/gmtsar/gmtsar
+        #BRANCH=master
+
+        # prepare system
+        #sudo apt update
+        #sudo apt -y upgrade
+        #sudo apt install -y locales
+        # Uncomment en_US.UTF-8 for inclusion in generation
+        #sudo sed -i 's/^# *\(en_US.UTF-8\)/\1/' /etc/locale.gen
+        # Generate locale
+        #sudo locale-gen en_US.UTF-8
+        # Export env vars
+        #sudo echo "export LC_ALL=en_US.UTF-8" >> /root/.bashrc
+        #sudo echo "export LANG=en_US.UTF-8" >> /root/.bashrc
+        #sudo echo "export LANGUAGE=en_US.UTF-8" >> /root/.bashrc
+
+        # https://github.com/gmtsar/gmtsar/wiki/GMTSAR-Wiki-Page
+        sudo apt install -y csh subversion autoconf libtiff5-dev libhdf5-dev wget
+        sudo apt install -y liblapack-dev
+        sudo apt install -y gfortran
+        sudo apt install -y g++
+        sudo apt install -y libgmt-dev
+        sudo apt install -y gmt-dcw gmt-gshhg
+        # gmt-gshhg-full should be installed automatically (it is required to use GMTSAR landmask)
+        sudo apt install -y gmt
+        # add missed package
+        sudo apt install -y make
+    - name: Compile GMTSAR
+      run: |
+        autoconf
+        ./configure --with-orbits-dir=/tmp
+        make
+        make install
+        # check installation
+        export PATH=$PATH:/home/runner/work/gmtsar/gmtsar/bin
+        echo "Start gmtsar_sharedir.csh"
+        gmtsar_sharedir.csh
+    - name: Install PyGMTSAR
+      run: |
+        #export PATH=$PATH:/home/runner/work/gmtsar/gmtsar/bin
+        # upgrade package to resolve dependencies
+        pip3 install click --upgrade
+        pip3 install -e ./pygmtsar/
+    - name: Download dataset
+      working-directory: tests
+      run: |
+        wget -c http://topex.ucsd.edu/gmtsar/tar/S1A_Stack_CPGF_T173.tar.gz
+        tar xvzf S1A_Stack_CPGF_T173.tar.gz -C .
+        rm -f S1A_Stack_CPGF_T173.tar.gz
+    - name: Run test
+      working-directory: tests
+      run: |
+        export PATH=$PATH:/home/runner/work/gmtsar/gmtsar/bin
+        #echo "Start gmtsar_sharedir.csh"
+        #gmtsar_sharedir.csh
+        #echo "Start S1A_Stack_CPGF_T173.py"
+        python3 ./S1A_Stack_CPGF_T173.py
