### 2022-09-18 3968b51
Clean config
