### 2022-09-18 1efe12d
Add CI and PyPi badges to documentation page
