+++ README.md
@@ -3,17 +3,13 @@
 ![Linux tests](https://github.com/mobigroup/gmtsar/actions/workflows/ubuntu.yml/badge.svg)
 [![Available on pypi](https://img.shields.io/pypi/v/pygmtsar.svg)](https://pypi.python.org/pypi/pygmtsar/)
 
-Note: this repository based on forked original GMTSAR plus my patches, shell scripts for SBAS pipeline on cloud hosts and Python codes. I commit my changes to GMTSAR and so it's possible to install original GMTSAR master branch following the upstream installation instructions and PyGMTSAR Python package via PIP. To simplify the process the Live example notebooks for Google Colab below automatically install all the software. MacOS and Linux Debian (Ubuntu) are my prefered OSes and I support both of them, see Debian cloud hosts init scripts in gmtsar/sh directory.
+This repository based on forked original GMTSAR and extended by my patches to binary tools and Python library PyGMTSAR. I commit my changes to binary tools to GMTSAR upstream and so it's possible to use the original GMTSAR master branch installation plus PyGMTSAR Python package via PIP. The project documentation including installation instructions available by the link: https://mobigroup.github.io/gmtsar/
 
-The goal of the project is easy and fast satellite interferometry (InSAR) processing everywhere as on localhost as on cloud environments like to Google Cloud VM and AI Notebooks and Amazon EC2 and on free of charge cloud environment Google Colab. GMTSAR binary command line tools are used under the hood but all GMTSAR scripts and GMT command replaced by Python code using modern and robust algorithms. By my opinion, GMTSAR is great project and has only one big problem - it is not user friendly. I build rich Python API instead of GMTSAR CSH scripts and GMT toolkit calls and by this way we have 3-10 times faster processing and more accurate results. How is it possible? The key is parallelization and algorithms - all the scenes and subswaths processing using all available processor cores and SNAPHU unwrapping configurations allow to use all the cores as for a single interferogram unwrapping as for multiple ones. Also, slow GMTSAR direct and inverse geocoding is replaced by fast matrix geocoding transformation and DEM processing using smoth cubic spline with smooth derivative instead of inaccurate and slow GMT tension surfaces (which derivative is not continuos on any scale so any numeric operations on it produces big artifacts). By this way, PyGMTSAR = GMTSAR - GMT + Python. And a couple of words about the really popular question - would be Python fast enough to use it for intensive numerical computations? The answer is exactly "yes" because basic scientific Python libraries coded using compiled C and Fortran languages and use as code parallelization as vectorization. Actually, GMTSAR binaries don't use any of these and Python processing can be significantly faster. Be careful Jupyter notebooks use lot of RAM for optional interactive maps (by default enabled on MacOS only where 16 GB RAM is enough). To limit RAM consuming use PyGMTSAR in command-line scripts still having progress indicators and other advantages.
-
-PyGMTSAR automatically downloads Sentinel-1 orbit files and SRTM DEM (and converts it to ellispoidal heights using EGM96 model) and even Sentinel-1 raw scenes using Alaska Satellite Facility (ASF) datastore.
-
-The project documentation including installation instructions available by the link: https://mobigroup.github.io/gmtsar/
+The goal of the project is easy and fast satellite interferometry (InSAR) processing everywhere as on localhost as on cloud environments like to Google Cloud VM and AI Notebooks and Amazon EC2 and on free of charge cloud environment Google Colab. I built the rich Python API and we have 3-10 times faster processing and more accurate results comparing to GMTSAR due to massive parallelization and better algorithms.
 
 ### Live Examples on Google Colab - just click on the links to run the processing in your own browser without any software installation
 
[…]
