### 2022-09-18 eccf727
Add initial workflow for MacOS
