+++ .github/workflows/macos.yml
@@ -1,0 +1,56 @@
+# This workflow will install Python dependencies, run tests and lint with a single version of Python
+# For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
+
+name: MacOS tests
+
+on:
+  push:
+    branches: [ "master" ]
+  pull_request:
+    branches: [ "master" ]
+
+permissions:
+  contents: read
+
+jobs:
+  S1A_Stack_CPGF_T173:
+
+    runs-on: macos-latest
+
+    steps:
+    - uses: actions/checkout@v3
+    - name: Set up Python 3.10
+      uses: actions/setup-python@v3
+      with:
+        python-version: "3.10"
+    - name: Install system dependencies
+      run: |
+        uname -a
+        # prepare system
+        brew install wget libtiff hdf5 gmt ghostscript autoconf
+    - name: Compile GMTSAR
+      run: |
+        autoconf
+        ./configure --with-orbits-dir=/tmp
+        make
+        make install
+        # check installation
+        export PATH=$PATH:/home/runner/work/gmtsar/gmtsar/bin
+        echo "Start gmtsar_sharedir.csh"
+        gmtsar_sharedir.csh
+    - name: Install PyGMTSAR
+      run: |
+        # upgrade package to resolve dependencies
+        pip3 install click --upgrade
+        pip3 install -e ./pygmtsar/
+    - name: Download dataset
+      working-directory: tests
+      run: |
+        wget -c http://topex.ucsd.edu/gmtsar/tar/S1A_Stack_CPGF_T173.tar.gz
+        tar xvzf S1A_Stack_CPGF_T173.tar.gz -C .
+        rm -f S1A_Stack_CPGF_T173.tar.gz
+    - name: Run test
+      working-directory: tests
+      run: |
+        export PATH=$PATH:/home/runner/work/gmtsar/gmtsar/bin
+        python3 ./S1A_Stack_CPGF_T173.py
