### 2022-09-18 ac08e41
Add MacOS CI badge
