+++ docs/README.md
@@ -1,5 +1,7 @@
-![Linux tests](https://github.com/mobigroup/gmtsar/actions/workflows/ubuntu.yml/badge.svg)
+![MacOS tests](https://github.com/mobigroup/gmtsar/actions/workflows/macos.yml/badge.svg)
+![Ubuntu tests](https://github.com/mobigroup/gmtsar/actions/workflows/ubuntu.yml/badge.svg)
 [![Available on pypi](https://img.shields.io/pypi/v/pygmtsar.svg)](https://pypi.python.org/pypi/pygmtsar/)
+![Build docs](https://github.com/mobigroup/gmtsar/actions/workflows/pages/pages-build-deployment/badge.svg)
 
 ## About
 
