+++ .github/workflows/ubuntu-test.yml
@@ -93,4 +93,4 @@ jobs:
       working-directory: tests
       run: |
         export PATH=$PATH:/home/runner/work/gmtsar/gmtsar/bin
-        #python3 ./S1A_2016_Kumamoto_Earthquake.py
+        python3 ./S1A_2016_Kumamoto_Earthquake.py
