### 2022-09-19 16f38f8
Add new CI test to Ubuntu workflow
