### 2022-09-19 700f7b0
Create CODE_OF_CONDUCT.md
