### 2022-09-19 4d0afa2
Fix progress bars for console run
