+++ pygmtsar/pygmtsar/SBAS.py
@@ -483,8 +483,7 @@ class SBAS:
         import elevation
         import os
         import subprocess
-        #from tqdm import tqdm
-        from tqdm import notebook
+        from tqdm.auto import tqdm
         import joblib
         # 0.000833333333333 cell size for SRTM3 90m
         # 0.000277777777778 cell size for SRTM1 30m
@@ -521,7 +520,7 @@ class SBAS:
                             margin=str(buffer_degrees),
                             output=os.path.realpath(tif_filename))
             elevation.clean()
-        with self.tqdm_joblib(notebook.tqdm(desc='Downloading', total=1)) as progress_bar:
+        with self.tqdm_joblib(tqdm(desc='Downloading', total=1)) as progress_bar:
             _ = joblib.Parallel(n_jobs=1)(joblib.delayed(func)() for i in [1])
 
         #if not os.path.exists(grd_filename):
@@ -556,8 +555,7 @@ class SBAS:
         from pygmtsar import PRM
         import os
         import subprocess
-        #from tqdm import tqdm
-        from tqdm import notebook
+        from tqdm.auto import tqdm
         # 0.000833333333333 cell size for SRTM3 90m
         # 0.000277777777778 cell size for SRTM1 30m
         scale = 0.000833333333333/90
@@ -604,7 +602,7 @@ class SBAS:
                 print (cmd.stdout)
 
         # use GMT commands pipeline to download and preprocess the DEM
-        with notebook.tqdm(desc='DEM Downloading', total=1) as pbar:
+        with tqdm(desc='DEM Downloading', total=1) as pbar:
             # get srtm data
             argv = ['gmt', 'grdcut', gridname, f'-R{minx}/{maxx}/{miny}/{maxy}', f'-G{ortho_filename}', '-V']
             run_cmd(argv)
@@ -630,8 +628,7 @@ class SBAS:
         """
         import os
         import subprocess
-        #from tqdm import tqdm
-        from tqdm import notebook
+        from tqdm.auto import tqdm
 
         if self.landmask_filename is not None:
             print ('NOTE: landmask exists, ignore the command. Use SBAS.set_landmask(None) to allow new landmask downloading')
@@ -660,7 +657,7 @@ class SBAS:
                 print (cmd.stdout)
 
         # use GMT commands pipeline to download and preprocess the DEM
-        with notebook.tqdm(desc='Landmask Downloading', total=1) as pbar:
+        with tqdm(desc='Landmask Downloading', total=1) as pbar:
             argv = ['gmt', 'grdlandmask', f'-G{landmask_filename}',
                     f'-R{llmin}/{llmax}/{ltmin}/{ltmax}', f'-I{scale}', '-V', '-NNaN/1', '-Df']
             run_cmd(argv)
@@ -974,8 +971,7 @@ class SBAS:
         return df
 
     def reframe_parallel(self, dates=None, n_jobs=-1, **kwargs):
-        #from tqdm import tqdm
-        from tqdm import notebook
+        from tqdm.auto import tqdm
         import joblib
         import pandas as pd
 
@@ -989,7 +985,7 @@ class SBAS:
         subswaths = self.get_subswaths()
 
         # process all the scenes
-        with self.tqdm_joblib(notebook.tqdm(desc='Reframing', total=len(dates)*len(subswaths))) as progress_bar:
+        with self.tqdm_joblib(tqdm(desc='Reframing', total=len(dates)*len(subswaths))) as progress_bar:
             records = joblib.Parallel(n_jobs=n_jobs)(joblib.delayed(self.reframe)(subswath, date, **kwargs) \
                                                      for date in dates for subswath in subswaths)
 
[…]
