### 2022-09-19 3e14f62
Use old backend to download DEM in CI test script
