### 2022-09-19 1bd815d
Fix for the recent script
