### 2022-09-19 4104ae0
Try to use SRTM3 vs SRTM1 DEM to resolve gmt downloading timeout on CI servers
