### 2022-09-19 3180e07
Add new CI Ubuntu test
