+++ .github/workflows/ubuntu.yml
@@ -149,3 +149,85 @@ jobs:
       run: |
         export PATH=$PATH:/home/runner/work/gmtsar/gmtsar/bin
         python3 ./S1A_2016_Kumamoto_Earthquake.py
+
+  S1AB_2021_Crete_Earthquake:
+
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v3
+    - name: Set up Python 3.10
+      uses: actions/setup-python@v3
+      with:
+        python-version: "3.10"
+    - name: Install system dependencies
+      run: |
+        uname -a
+        # prepare system
+        sudo apt update
+        # https://github.com/gmtsar/gmtsar/wiki/GMTSAR-Wiki-Page
+        sudo apt install -y csh subversion autoconf libtiff5-dev libhdf5-dev wget
+        sudo apt install -y liblapack-dev
+        sudo apt install -y gfortran
+        sudo apt install -y g++
+        sudo apt install -y libgmt-dev
+        sudo apt install -y gmt-dcw gmt-gshhg
+        # gmt-gshhg-full should be installed automatically (it is required to use GMTSAR landmask)
+        sudo apt install -y gmt gmt-gshhg-full
+        # add missed package
+        sudo apt install -y make
+    - name: Compile GMTSAR
+      run: |
+        autoconf
+        ./configure --with-orbits-dir=/tmp
+        make
+        make install
+        # check installation
+        export PATH=$PATH:/home/runner/work/gmtsar/gmtsar/bin
+        echo "Start gmtsar_sharedir.csh"
+        gmtsar_sharedir.csh
+    - name: Install PyGMTSAR
+      run: |
+        # upgrade package to resolve dependencies
+        pip3 install click --upgrade
+        pip3 install -e ./pygmtsar/
+    - name: Download dataset
+      working-directory: tests
+      run: |
+        # download iCloud datasets
+        cat > icloud_urls.txt <<- END
+        https://www.icloud.com/iclouddrive/060EcVsMZF5o5nDoyAxHZ-JgQ#S1A_IW_SLC__1SDV_20210929T162245_20210929T162312_039899_04B89A_651D.zip
+        https://www.icloud.com/iclouddrive/0bcQQkK2QDSiX76_y1O6zfd_Q#S1B_IW_SLC__1SDV_20210923T162204_20210923T162231_028828_0370C4_681D.zip
+        https://www.icloud.com/iclouddrive/085_X__5ZszrqcnP--SGCH-tw#S1AB_Crete_Earthquake_vs_geObsevatory.jpg
+        END
+        cut -f5 -d '/' icloud_urls.txt | cut -f1 -d '#' | sed -E 's/(.*)/{"shortGUIDs":[{"value":"\1"}]}/' > requests.txt
+        cut -f5 -d '/' icloud_urls.txt | cut -f2 -d '#' > fnames.txt
+        paste -d ' ' requests.txt fnames.txt  | while read LINE
+        do
+            request=$(echo "$LINE"| cut -d ' ' -f1)
+            fname=$(echo "$LINE"| cut -d ' ' -f2)
+            echo "request: $request fname: $fname"
+            url=$(curl -s 'https://ckdatabasews.icloud.com/database/1/com.apple.cloudkit/production/public/records/resolve' \
+            --data-raw "${request}" --compressed | jq -r '.results[0].rootRecord.fields.fileContent.value.downloadURL')
+            curl -so "${fname}" "${url}"
+        done
+        # unpack to data directory
+        mkdir -p data
[…]
