### 2022-09-19 95c69fa
Update PyPi package version as all the CI tests passed
