### 2022-09-19 #60 @lwasser
👋 hi @mobigroup thank you for this submission. I will get back to you with next steps. In the meantime can you explain what the GMTSAR binary tools are / entail? does this package wrap around them to perform tasks on sentinel data? or is it an API wrapper to download the data? Many thanks for any clarification.
