### 2022-09-20 #60 @mobigroup
Hi @lwasser GMTSAR is a well-known satellite interferometry (InSAR) system based on GMT tools for geospatial data processing, see details on the project page: https://topex.ucsd.edu/gmtsar/ You’d use GMTSAR for the complete interferometry processing but it doesn’t have any user interface because it’s a set of linux console binaries and shell scripts. Every processing requires lots of manual work even when you are a linux Jedi. Also, there are some issues in the processing like to outdated interpolation methods. I use GMTSAR tools but exclude all the GMT tools for the new generation of a satellite interferometry system. I’m sorry if it sounds not so easy to understand because satellite interferometry system includes hundreds of different tools and we try to reuse some existing ones when it’s possible and add dozens of new ones (because building it from the scratch requires decades of work).

I’m a single developer of PyGMTSAR and there are no other package maintainers. PyGMTSAR GitHub repository includes some 3rd party open source software like to GMTSAR and SNAPHU (satellite interferometry phase unwrapper) and I work on GMTSAR codes too (and send the patches to it’s GitHub) to adopt the required tools to call them from PyGMTSAR. PyGMTSAR python library is located in the project subdirectory (as GMTSAR and SNAPHU codes) and it’s available as PyPi package: https://pypi.org/project/pygmtsar/ Actually, I work on the two projects at the same time: GMTSAR and PyGMTSAR where GMTSAR has lots of developers and PyGMTSAR is my own project.
