### 2022-09-20 1955a27
Simplify PyPI tests to prevent network timeouts
