### 2022-09-20 7f2f967
Add PyPI tests badge
