+++ .github/workflows/pypi.yml
@@ -59,6 +59,7 @@ jobs:
       with:
         path: tests/S1A_Stack_CPGF_T173.tar.gz
         key: S1A_Stack_CPGF_T173.tar.gz
+        restore-keys: S1A_Stack_CPGF_T173.tar.gz
     - name: Download dataset
       working-directory: tests
       run: |
