### 2022-09-20 4bf81f8
Reorder CI tests badges
