### 2022-09-20 c14c9f3
Add caching for dataset S1A_Stack_CPGF_T173.tar.gz for all the tests
