### 2022-09-21 cd97db3
Add CI matrix to test on Python 3.7 ... 3.10 and MacOS 11, 12
