### 2022-09-21 865526a
Add matrix of OSes and Python versions for CI Ubuntu workflow and use only one SBAS test for all the jobs
