+++ .github/workflows/ubuntu.yml
@@ -15,14 +15,17 @@ permissions:
 jobs:
   S1A_Stack_CPGF_T173:
 
-    runs-on: ubuntu-latest
-
+    strategy:
+      matrix:
+        os: [ubuntu-22.04, ubuntu-20.04, ubuntu-18.04]
+        python-version: ["3.7", "3.8", "3.9", "3.10", "3.11"]
+    runs-on: ${{ matrix.os }}
     steps:
     - uses: actions/checkout@v3
-    - name: Set up Python 3.10
+    - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v3
       with:
-        python-version: "3.10"
+        python-version: ${{ matrix.python-version }}
     - name: Install system dependencies
       run: |
         uname -a
@@ -59,7 +62,7 @@ jobs:
       with:
         path: tests/S1A_Stack_CPGF_T173.tar.gz
         key: S1A_Stack_CPGF_T173.tar.gz
-        restore-keys: S1A_Stack_CPGF_T173.tar.gz
+        restore-keys: S1A_Stack_CPGF_T173
     - name: Download dataset
       working-directory: tests
       run: |
@@ -71,169 +74,3 @@ jobs:
       run: |
         export PATH=$PATH:/home/runner/work/gmtsar/gmtsar/bin
         python3 ./S1A_Stack_CPGF_T173.py
-
-  S1A_2016_Kumamoto_Earthquake:
-
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v3
-    - name: Set up Python 3.10
-      uses: actions/setup-python@v3
-      with:
-        python-version: "3.10"
-    - name: Install system dependencies
-      run: |
-        uname -a
-        # prepare system
-        sudo apt update
-        # https://github.com/gmtsar/gmtsar/wiki/GMTSAR-Wiki-Page
-        sudo apt install -y csh subversion autoconf libtiff5-dev libhdf5-dev wget
-        sudo apt install -y liblapack-dev
-        sudo apt install -y gfortran
-        sudo apt install -y g++
-        sudo apt install -y libgmt-dev
-        sudo apt install -y gmt-dcw gmt-gshhg
-        # gmt-gshhg-full should be installed automatically (it is required to use GMTSAR landmask)
-        sudo apt install -y gmt
-        # add missed package
-        sudo apt install -y make
-        # for old DEM downloading backend
-        sudo apt install -y gdal-bin
-    - name: Compile GMTSAR
-      run: |
-        autoconf
-        ./configure --with-orbits-dir=/tmp
-        make
-        make install
-        # check installation
-        export PATH=$PATH:/home/runner/work/gmtsar/gmtsar/bin
-        echo "Start gmtsar_sharedir.csh"
-        gmtsar_sharedir.csh
-    - name: Install PyGMTSAR
-      run: |
-        # upgrade package to resolve dependencies
-        pip3 install click --upgrade
-        pip3 install -e ./pygmtsar/
-    - name: Download dataset
-      working-directory: tests
-      run: |
-        # download iCloud datasets
-        cat > icloud_urls.txt <<- END
-        https://www.icloud.com/iclouddrive/04946wsi9N5g0W6beTl_KUvQg#S1A_IW_SLC__1SSV_20160408T091355_20160408T091430_010728_01001F_83EB.zip
[…]
