### 2022-09-21 b579a45
Remove CI Python 3.11 as unsupported
