### 2022-09-21 09f50b4
Add double quotas forUbuntu versions in CI test
