### 2022-09-21 66aae5b
Remove CI tests on Ubuntu 22.04 due to GMTSAR compilation problem
