### 2022-09-21 4bb3b85
Allow to continue one some of tests failes for CI MacOS
