### 2022-09-21 ae4a85f
Move 3 different tests to PyPI CI workflow and run them ofor Python 3.7 and 3.10
