### 2022-09-21 2fe2e35
Fix for double call of applied functions in SBAS.open_grids()
