### 2022-09-21 87c3b7f
Add gdal package installation to use old DEM downloading backend for CI test
