+++ .github/workflows/pypi.yml
@@ -193,6 +193,8 @@ jobs:
         sudo apt install -y gmt gmt-gshhg-full
         # add missed package
         sudo apt install -y make
+        # for old DEM downloading backend
+        sudo apt install -y gdal-bin
     - name: Compile GMTSAR
       run: |
         autoconf
