### 2022-09-22 d64362b
Fix detrend function to work in two modes: interactive to return results and not interactive to save the results on disk
