### 2022-09-22 12fed4a
Add interactive/not interactive mode to SBAS.unwrap() function
