### 2022-09-22 2ef683f
Fix connected components processing in SBAS.unwrap() function
