### 2022-09-23 8d5d53d
Fix GMTSAR compilation on Ubuntu 22.04. Add CI test configuration and update documentation.
