### 2022-09-24 8c6d124
Enhance SBAS.backup() function and add test for it.
