### 2022-09-24 cd3e900
Drop dataset removing command in MacOS tests
