### 2022-09-25 798c9c4
Drop all my scripts in gmtsar directory to revert it to the original upstream state. Move notebooks subdirectory up. Use PyGMTSAR instead any of my Python scripts or upstream shell scripts.
