+++ gmtsar/py/run_pytest.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-python3.9 -m pytest *.py
