### 2022-09-25 6d12a35
Add error message for invalid pins order
