### 2022-09-26 d7421a8
Add CI test for GMTSAR itself (#447)
