### 2022-09-27 674452d
Add debug option to SBAS.backup() function and clean paths in SBAS dataframe when the files moved (copy=False).
