+++ tests/S1A_Stack_CPGF_T173.py
@@ -73,10 +73,6 @@ def interp(da):
 dates = np.unique(pairs.values.flatten())
 disps = sbas.open_grids(dates, 'disp', func=[sbas.los_displacement_mm, interp])
 print (sbas.cropna(sbas.intf_ra2ll(disps.cumsum('date').sel(x=slice(9000,12000), y=slice(1800,2700)))))
-# dump current state
-sbas.dump()
-# backup
-sbas.backup('backup')
 
 # test results
 # https://github.com/mobigroup/gmtsar/issues/2
@@ -93,19 +89,32 @@ xr.testing.assert_allclose(
     sbas.intf_ra2ll(sbas.open_grids(None, 'vel', add_subswath=False))
 )
 
-# test dump and dump backup
+# dump current state
+sbas.dump()
 assert os.path.exists(os.path.join(WORKDIR,   'SBAS.pickle')), 'State file not found'
-assert os.path.exists(os.path.join(BACKUPDIR, 'SBAS.pickle')), 'Backup state file not found'
 
-# test backup
+# backup (copy)
+sbas.backup('backup', copy=True)
+assert os.path.exists(os.path.join(BACKUPDIR, 'SBAS.pickle')), 'Backup state file not found'
 scenes = sbas.df.datapath.apply(lambda fn: os.path.splitext(os.path.split(fn)[-1])[0]).values
 for scene in scenes:
     for ext in ['.tiff', '.xml']:
         assert os.path.exists(os.path.join(BACKUPDIR,   scene+ext)), 'Backup file not found'
         assert os.path.exists(os.path.join(DATADIR,     scene+ext)), 'Original file not found'
         assert not os.path.exists(os.path.join(WORKDIR, scene+ext)), 'Excessive file found'
-
 orbits = sbas.df.orbitpath.apply(lambda fn: os.path.split(fn)[-1]).values
+for orbit in orbits:
+    assert os.path.exists(os.path.join(BACKUPDIR,   orbit)), 'Backup file not found'
+    assert os.path.exists(os.path.join(DATADIR,     orbit)), 'Original file not found'
+    assert not os.path.exists(os.path.join(WORKDIR, orbit)), 'Excessive file found'
+
+# backup (move only local files)
+sbas.backup('backup', copy=False)
+for scene in scenes:
+    for ext in ['.tiff', '.xml']:
+        assert os.path.exists(os.path.join(BACKUPDIR,   scene+ext)), 'Backup file not found'
+        assert os.path.exists(os.path.join(DATADIR,     scene+ext)), 'Original file not found'
+        assert not os.path.exists(os.path.join(WORKDIR, scene+ext)), 'Excessive file found'
 for orbit in orbits:
     assert os.path.exists(os.path.join(BACKUPDIR,   orbit)), 'Backup file not found'
     assert os.path.exists(os.path.join(DATADIR,     orbit)), 'Original file not found'
