### 2022-09-27 ccc2ffd
Update GMTSAR CI test to fix 'tests' directory using which is missed in the upstream
