+++ .github/workflows/gmtsar.yml
@@ -51,16 +51,14 @@ jobs:
     - name: Cache dataset
       uses: actions/cache@v3
       with:
-        path: tests/S1A_Stack_CPGF_T173.tar.gz
+        path: S1A_Stack_CPGF_T173.tar.gz
         key: S1A_Stack_CPGF_T173.tar.gz
         restore-keys: S1A_Stack_CPGF_T173
     - name: Download dataset
-      working-directory: tests
       run: |
         wget -c http://topex.ucsd.edu/gmtsar/tar/S1A_Stack_CPGF_T173.tar.gz
         tar xvzf S1A_Stack_CPGF_T173.tar.gz -C .
     - name: Run test
-      working-directory: tests
       run: |
         export PATH=$PATH:/home/runner/work/gmtsar/gmtsar/bin
         ./README_prep.txt
@@ -94,16 +92,14 @@ jobs:
     - name: Cache dataset
       uses: actions/cache@v3
       with:
-        path: tests/S1A_Stack_CPGF_T173.tar.gz
+        path: S1A_Stack_CPGF_T173.tar.gz
         key: S1A_Stack_CPGF_T173.tar.gz
         restore-keys: S1A_Stack_CPGF_T173
     - name: Download dataset
-      working-directory: tests
       run: |
         wget -c http://topex.ucsd.edu/gmtsar/tar/S1A_Stack_CPGF_T173.tar.gz
         tar xvzf S1A_Stack_CPGF_T173.tar.gz -C .
     - name: Run test
-      working-directory: tests
       run: |
         export PATH=$PATH:/Users/runner/work/gmtsar/gmtsar/bin
         ./README_prep.txt
