### 2022-09-27 6b3c9f1
Add ghostscript package installation for GMTSAR Ubuntu CI script
