+++ .github/workflows/gmtsar.yml
@@ -35,7 +35,7 @@ jobs:
         sudo apt install -y libgmt-dev
         sudo apt install -y gmt-dcw gmt-gshhg
         # gmt-gshhg-full should be installed automatically (it is required to use GMTSAR landmask)
-        sudo apt install -y gmt
+        sudo apt install -y gmt ghostscript
         # add missed package
         sudo apt install -y make
     - name: Compile GMTSAR
