### 2022-09-27 6cc2d61
Fix CI test (#451)
* Add ghostscript package installation for GMTSAR Ubuntu CI script

* Update GMTSAR CI test to fix 'tests' directory using which is missed in the upstream
