### 2022-09-27 9b3a5dd
fixing a bug in make_slc_gf3.c suggested by @zhangdabao2020
