+++ preproc/GF3_preproc/src/make_slc_gf3.c
@@ -119,8 +119,8 @@ int write_slc(TIFF *tif, FILE *slc, char *orbdir) {
 	// printf("%d %d \n",width,height);
 
 	buf = (uint16 *)_TIFFmalloc(TIFFScanlineSize(tif));
-	tmp = (short *)malloc(width * 2 * sizeof(short));
 	width = widthi - widthi % 4;
+	tmp = (short *)malloc(width * 2 * sizeof(short));
 	printf("Writing SLC..Image Size: %d X %d...\n", width, height);
 
     for (i = 0; i < height; i++) {
