+++ pygmtsar/pygmtsar/SBAS.py
@@ -1297,15 +1297,78 @@ class SBAS:
 
         return grids_ll
 
-    def topo_ra_parallel(self, n_jobs=-1, **kwargs):
+    def topo_ra_parallel(self, idec=2, jdec=2, n_jobs=-1, debug=False):
+        import numpy as np
+        import xarray as xr
+        from scipy.spatial import cKDTree
         from tqdm.auto import tqdm
         import joblib
-    
+        import os
+
+        def SAT_llt2rat(subswath, chunk):
+                return self.PRM(subswath).SAT_llt2rat(chunk, precise=1, binary=True)
+
+        # start a set of jobs together but not more than available cpu cores at once
+        if n_jobs == -1:
+            n_jobs = joblib.cpu_count()
+
+        # as defined in dem2topo_ra.csh for Sentinel-1
+        if debug:
+            print (f'DEBUG: Topography range and azimuth decimation: {idec}/{jdec}')
+
         # process all the subswaths
         subswaths = self.get_subswaths()
-        with self.tqdm_joblib(tqdm(desc='Transforming Coordinates', total=len(subswaths))) as progress_bar:
-            joblib.Parallel(n_jobs=n_jobs)(joblib.delayed(self.topo_ra)(subswath, **kwargs) \
-                                                     for subswath in subswaths)
+        for subswath in subswaths:
+            # process selected subswaths 1,2,3 or 12, 23, 123
+            #print ('subswath', subswath)
+            prefix = f'F{subswath}_'
+            trans_dat_file = os.path.join(self.basedir, prefix + 'trans.dat')
+            topo_ra_file = os.path.join(self.basedir, prefix + 'topo_ra.grd')
+
+            # cleanup before creating the new files
+            for filename in [trans_dat_file, topo_ra_file]:
+                if os.path.exists(filename):
+                    os.remove(filename)
+
+            XMAX, yvalid, num_patch = self.PRM(subswath).get('num_rng_bins', 'num_valid_az', 'num_patches')
+            YMAX = yvalid * num_patch
+            #print ('XMAX', XMAX, 'YMAX', YMAX)
+            # use center pixel GMT registration mode
+            rngs = np.arange(1, XMAX+1, idec)
+            azis = np.arange(1, YMAX+1, jdec)
+            grid_r, grid_a = np.meshgrid(rngs, azis)
+
+            # build trans.dat
+            dem = self.get_dem(subswath, geoloc=True)
+            # generated by llt_grid2rat (r a topo lon lat)"
+            lats, lons, z = xr.broadcast(dem.lat, dem.lon, dem)
+            dem_data = np.column_stack([lons.values.ravel(), lats.values.ravel(), z.values.ravel()])
+
+            dem_data_chunks = np.array_split(dem_data, n_jobs)
+            with self.tqdm_joblib(tqdm(desc=f'Topography Processing', total=n_jobs)) as progress_bar:
+                trans_dat_chunks = joblib.Parallel(n_jobs=n_jobs)(joblib.delayed(SAT_llt2rat)(subswath, dem_data_chunk) \
+                                               for dem_data_chunk in dem_data_chunks)
+            trans_dat = np.concatenate(trans_dat_chunks)
+            # save to binary file
+            with open(trans_dat_file, 'wb') as f:
+                f.write(trans_dat)
+
+            # find nearest DEM values for every radar grid point
+            ras = np.stack([grid_r.reshape(-1), grid_a.reshape(-1)], axis=1)
+            trans_ras = np.stack([trans_dat[:,0],trans_dat[:,1]], axis=1)
+            tree = cKDTree(trans_ras, compact_nodes=False, balanced_tree=False)
+            d, inds = tree.query(ras, k = 1, workers=n_jobs)
[…]
