+++ pygmtsar/pygmtsar/PRM.py
@@ -655,79 +655,6 @@ class PRM:
 
         return prm
 
-    # create replacement for trans.dat - file generated by llt_grid2rat (r a topo lon lat)"
-    def topo_ra(self, dem, trans_dat_tofile, topo_ra_tofile=None, method='cubic'):
-        import numpy as np
-        import xarray as xr
-        from scipy.interpolate import griddata
-        from scipy.ndimage.filters import gaussian_filter
-
-        XMAX, yvalid, num_patch = self.get('num_rng_bins', 'num_valid_az', 'num_patches')
-        YMAX = yvalid * num_patch
-        #print ('XMAX', XMAX, 'YMAX', YMAX)
-        # as defined in dem2topo_ra.csh for Sentinel-1
-        azi_dec = 2
-        rng_dec = 2
-        print (f'Topography range and azimuth decimation: {rng_dec}/{azi_dec}')
-
-        # use center pixel GMT registration mode 
-        rngs = np.arange(1, XMAX+1, rng_dec)
-        azis = np.arange(1, YMAX+1, azi_dec)
-        grid_r, grid_a = np.meshgrid(rngs, azis)
-        #print ('grid_r', grid_r.shape)
-
-        # build trans.dat
-        # generated by llt_grid2rat (r a topo lon lat)"
-        lats, lons, z = xr.broadcast(dem.lat, dem.lon, dem)
-        dem_data = np.column_stack([lons.values.ravel(), lats.values.ravel(), z.values.ravel()])
-        #print ('dem_data', dem_data.shape)
-        #if trans_dat_tofile is not None:
-        self.SAT_llt2rat(dem_data, tofile=trans_dat_tofile, precise=1, binary=True)
-
-        # build topo_ra
-        coords = np.fromfile(trans_dat_tofile, dtype=np.float64).reshape([-1,5])
-        # or return numpy array and save it to the trans_dat file later
-        #coords = self.SAT_llt2rat(dem_data, precise=1, binary=True)
-        if np.any(np.isnan(coords)):
-            raise Exception('Invalid trans.dat file with NaN values detected')
-
-        # also, we can use all cores for the calculation (and save 2-3 minutes from 15 in total) while the longest processing is below
-        #import joblib
-        #from joblib.externals import loky
-        # start a set of jobs together but not more than available cpu cores at once
-        #n_jobs = joblib.cpu_count()
-        #loky.get_reusable_executor(kill_workers=True).shutdown(wait=True)
-        #with joblib.parallel_backend('loky', n_jobs=n_jobs, inner_max_num_threads=1):
-        #    coords = joblib.Parallel()(joblib.delayed(self.SAT_llt2rat)(dem_data_chunk, precise=1, binary=True)
-        #                               for dem_data_chunk in np.array_split(dem_data, n_jobs))
-        #coords = np.concatenate(coords)
-        #if trans_dat_tofile is not None:
-        #    with open(trans_dat_tofile, 'wb') as f:
-        #        f.write(coords)
-
-        # this processing is not parallelized and it is time consuming
-        # TODO: plit to overlapping chunks and process them in parallel
-        # fill_value used to fill in outside of the convex hull of the input points (default is nan)
-        grid = griddata((coords[:,0], coords[:,1]), coords[:,2], (grid_r, grid_a), method=method, fill_value=0)
-        
-        # remove subpixel noise
-        grid = gaussian_filter(grid, 1.0, mode='constant', cval=0)
-        # wrap to dataarray
-        topo = xr.DataArray(np.flipud(grid), coords={'y': azis, 'x': rngs}, name='z')
-
-        if topo_ra_tofile is None:
-            if np.any(np.isnan(topo)):
[…]
