### 2022-09-29 c25da30
Add more debug outputs
