### 2022-09-29 #60 @lwasser
hi @mobigroup thank you for your patience with us as we stand up pyOpenSci. I think you have CI already so we can check that box. I am going to look for someone in the community to have a look at this package. We are just getting our editorial board setup and I am trying to create structure for the organization. I want to make sure that you have a good experience with us. Please know that this process may be a bit slower than usual as we get things together.

**One other question - are there any other packages that perform overlapping tasks to this package?**

Many thanks
