### 2022-09-29 #60 @mobigroup
Hi @lwasser thanks, I understand it’s not so easy to review complex software packages! And yes, I’ve added CI tests recently.

> One other question - are there any other packages that perform overlapping tasks to this package?

While there are dozens of satellite interferometry packages I don’t know any other available for everyone (you don’t need to write very complex configurations in many separate files and so on) and everywhere (as on local computer as on cloud servers and even on Google Colab) using Python scripts or Jupyter notebooks. Please follow the link to [live Google Colab notebook](https://colab.research.google.com/drive/1sljxm2jAMGXynq4EYam6Siz8OLcPLN0h?usp=sharing) to compute and plot the results (and edit the code and interactively view the changed results) in your browser without long hassles with software installation and other problems. At the same time, PyGMTSAR works fast using parallelization intensively and allows to process huge datasets on very common hardware like to Apple Air (using lazy computations provided by dask and xarray packages). You’d find the community on LinkedIn where I post the package updates and collect the feedback: https://www.linkedin.com/in/alexey-pechnikov/recent-activity/
