### 2022-09-30 b9e77e4
Emhance debug output for subswaths merging functions
