### 2022-09-30 5fa3eaf
Enhance progress indication for subswaths merging
