### 2022-09-30 9ca2f6a
Add function SBAS.pixel_decimator() to generate post-processing function for interferogram calculation. Updated and rename related SBAS.pixel_sie() function to process multiple subswaths. Update the documentation.
