### 2022-09-30 bf23610
Update CI test for the recent changes
