### 2022-09-30 698d012
Update CI test for the recent changes
