### 2022-09-30 a946b18
Enhance debug logging
