+++ pygmtsar/pygmtsar/SBAS.py
@@ -141,7 +141,7 @@ class SBAS:
         sigma_x = np.round(wavelength / dx)
     
         if debug:
-            print ('NOTE: Gaussian filtering sigma_y, sigma_x', sigma_y, sigma_x)
+            print ('DEBUG: Gaussian filtering sigma_y, sigma_x', sigma_y, sigma_x)
     
         # there is a compromise between accuracy and speed
         dec = 32
@@ -152,7 +152,7 @@ class SBAS:
         if approximate and dec_y > 1 and dec_x > 1:
             # filter decimated array for faster and less accurate processing
             if debug:
-                print ('NOTE: Gaussian filtering decimation dec_y, dec_x', dec_y, dec_x)
+                print ('DEBUG: Gaussian filtering decimation dec_y, dec_x', dec_y, dec_x)
         
             # decimate and filter
             da_dec = da.coarsen({'y': dec_y, 'x': dec_x}, boundary='pad').mean()
@@ -563,13 +563,13 @@ class SBAS:
                 '-ot', 'Float32', '-of', 'NetCDF',
                 'DEM_EGM96.tif', 'DEM_WGS84.nc']
         if debug:
-            print ('argv', argv)
+            print ('DEBUG: argv', argv)
         p = subprocess.Popen(argv, stdout=subprocess.PIPE, stderr=subprocess.PIPE, cwd=self.basedir)
         stdout_data, stderr_data = p.communicate()
-        if len(stderr_data) > 0:
-            print ('download_dem', stderr_data.decode('ascii'))
+        if len(stderr_data) > 0 and debug:
+            print ('DEBUG: download_dem', stderr_data.decode('ascii'))
         if len(stdout_data) > 0 and debug:
-            print ('download_dem', stdout_data.decode('ascii'))
+            print ('DEBUG: download_dem', stdout_data.decode('ascii'))
 
         self.dem_filename = grd_filename
 
@@ -623,7 +623,7 @@ class SBAS:
         # helper function to run external commands
         def run_cmd(argv):
             if debug:
-                print ('argv', argv)
+                print ('DEBUG: argv', argv)
             cmd = subprocess.run(argv, capture_output=True)
             if cmd.returncode != 0:
                 print (cmd.stderr)
@@ -677,7 +677,7 @@ class SBAS:
         # helper function to run external commands
         def run_cmd(argv):
             if debug:
-                print ('argv', argv)
+                print ('DEBUG: argv', argv)
             cmd = subprocess.run(argv, capture_output=True)
             if cmd.returncode != 0:
                 print (cmd.stderr)
@@ -1513,13 +1513,13 @@ class SBAS:
 
         argv = ['assemble_tops', azi_1, azi_2] + datapaths + [stem]
         if debug:
-            print ('argv', argv)
+            print ('DEBUG: argv', argv)
         p = subprocess.Popen(argv, stdout=subprocess.PIPE, stderr=subprocess.PIPE, cwd=self.basedir)
         stdout_data, stderr_data = p.communicate()
         if len(stderr_data) > 0 and debug:
-            print ('assemble_tops', stderr_data.decode('ascii'))
+            print ('DEBUG: assemble_tops', stderr_data.decode('ascii'))
         if len(stdout_data) > 0 and debug:
-            print ('assemble_tops', stdout_data.decode('ascii'))
+            print ('DEBUG: assemble_tops', stdout_data.decode('ascii'))
 
         return
 
@@ -1536,13 +1536,13 @@ class SBAS:
 
         argv = ['ext_orb_s1a', f'{stem}.PRM', orbit, stem]
         if debug:
-            print ('argv', argv)
+            print ('DEBUG: argv', argv)
[…]
