### 2022-10-01 a1f0aa5
Add SBAS.restore() command to CI test
