### 2022-10-02 e49bcd0
Fix fo #4. When only a single DEM point processed by SAT_llt2rat we need to reshape the output.
