### 2022-10-02 eb9f67d
Fix pinning when master image should be stitched too
