### 2022-10-02 4fdb1d9
Enhance scenes assemble processing and debug logging
