### 2022-10-02 96843c1
Allow to load grids with a different dimension sizes (these are very close but not equal sometimes with difference in 1-2 pixels).
