### 2022-10-02 8ae8d98
Revert check for pins inside master scene(s)
