### 2022-10-03 2308cd5
removed ALOS_baseline and ALOS_look from the Makefile
