### 2022-10-04 2f7a00b
Refactor code to use more lazy computations. Update documentation. Make satellite look vector calculation optional.
