+++ pygmtsar/pygmtsar/SBAS.py
@@ -91,17 +91,6 @@ class SBAS:
         from pygmtsar import PRM
         return PRM.nearest_grid(in_grid, search_radius_pixels)
 
-    @staticmethod
-    def flipud(da):
-        """
-        Flip dataset vertically required for topography grid because it stored flipped for GMTSAR compatibility reasons. 
-        """
-        import numpy as np
-        # prevent input variable modification
-        da = da.copy()
-        da.values = np.flipud(da.values)
-        return da
-
     def snaphu_config(self, defomax=0, **kwargs):
         return self.PRM().snaphu_config(defomax, **kwargs)
  
@@ -762,23 +751,23 @@ class SBAS:
     def get_dem(self, subswath=None, geoloc=False, buffer_degrees=0.02):
         import xarray as xr
         import os
-        
+
         if self.dem_filename is None:
             raise Exception('Set DEM first')
 
         # open DEM file and find the elevation variable
         # because sometimes grid includes 'crs' or other variables
-        dem = xr.open_dataset(self.dem_filename, engine=self.netcdf_engine)
-        assert 'lat' in dem.coords and 'lon' in dem.coords
+        dem = xr.open_dataset(self.dem_filename, engine=self.netcdf_engine, chunks=2*self.netcdf_chunksize)
+        assert 'lat' in dem.coords and 'lon' in dem.coords, 'DEM should be defined as lat,lon grid'
         # define latlon array
         z_array_name = [data_var for data_var in dem.data_vars if len(dem.data_vars[data_var].coords)==2]
         assert len(z_array_name) == 1
         # extract the array and fill missed values (mostly water surfaces)
         dem = dem[z_array_name[0]].fillna(0)
-        
+
         if geoloc is False:
             return dem
-        
+
         bounds = self.get_master(subswath).dissolve().envelope.bounds.values[0].round(3)
         #print ('xmin, xmax', xmin, xmax)
         return dem\
@@ -797,7 +786,7 @@ class SBAS:
         
         # open DEM file and find the elevation variable
         # because sometimes grid includes 'crs' or other variables
-        landmask = xr.open_dataset(self.landmask_filename, engine=self.netcdf_engine)
+        landmask = xr.open_dataset(self.landmask_filename, engine=self.netcdf_engine, chunks=2*self.netcdf_chunksize)
         assert 'lat' in landmask.coords and 'lon' in landmask.coords
         # define latlon array
         z_array_name = [data_var for data_var in landmask.data_vars if len(landmask.data_vars[data_var].coords)==2]
@@ -1055,8 +1044,8 @@ class SBAS:
         intf_ra2ll_file = os.path.join(self.basedir, f'F{subswath}_intf_ra2ll.grd')
         intf_ll2ra_file = os.path.join(self.basedir, f'F{subswath}_intf_ll2ra.grd')
 
-        matrix_ra2ll = xr.open_dataarray(intf_ra2ll_file, engine=self.netcdf_engine)
-        matrix_ll2ra = xr.open_dataarray(intf_ll2ra_file, engine=self.netcdf_engine)
+        matrix_ra2ll = xr.open_dataarray(intf_ra2ll_file, engine=self.netcdf_engine, chunks=2*self.netcdf_chunksize)
+        matrix_ll2ra = xr.open_dataarray(intf_ll2ra_file, engine=self.netcdf_engine, chunks=2*self.netcdf_chunksize)
 
         # conversion works for a different 1st grid dimension size
         def ra2ll(grid):
@@ -1103,17 +1092,16 @@ class SBAS:
         # sometimes interferogram grids are different for one azimuth line so use the largest grid
         intf_grid = intf_grids.min('pair')
 
[…]
