### 2022-10-04 c1e68e0
Fix for updated before PRM.SAT_look() function which returns 1D array for a single pixel vs 2D array for multiple ones.
