### 2022-10-05 ace7eab
Fix auto-pinning following the recent changes
