### 2022-10-05 d3d52cc
Update CI test scripts following the recent changes
