+++ tests/S1AB_2021_Crete_Earthquake.py
@@ -54,6 +54,7 @@ sbas.unwrap_parallel(pairs, threshold=CORRLIMIT, mask=landmask_ra, func=cleaner)
 print (sbas.open_grids(pairs, 'unwrap', geocode=True)[0])
 print (sbas.open_grids(pairs, 'unwrap', func=sbas.los_displacement_mm, geocode=True)[0])
 # Incidence angle
+sbas.sat_look_parallel()
 print (sbas.incidence_angle())
 # Displacements
 print (sbas.open_grids(pairs, 'unwrap', func=sbas.los_displacement_mm, geocode=True)[0])
