### 2022-10-05 039fe31
Fix a problem from importlib-metadata library released 2019-10-02 on Python 3.7 only, see https://stackoverflow.com/questions/73929564/entrypoints-object-has-no-attribute-get-digital-ocean
