+++ pygmtsar/setup.py
@@ -9,7 +9,7 @@ long_description = (this_directory / "README.md").read_text()
 
 setup(
     name='pygmtsar',
-    version='2022.10.2',
+    version='2022.10.5',
     description='PyGMTSAR (Python GMTSAR) - Easy and Fast Satellite Interferometry For Everyone',
     long_description=long_description,
     long_description_content_type='text/markdown',
@@ -19,6 +19,7 @@ setup(
     license='GPL-3.0',
     packages=['pygmtsar'],
     install_requires=['xarray>=0.19.0',
+                      'importlib-metadata<5.0',
                       'numpy',
                       'pandas',
                       'geopandas',
