+++ pygmtsar/pygmtsar/SBAS.py
@@ -757,7 +757,7 @@ class SBAS:
 
         # open DEM file and find the elevation variable
         # because sometimes grid includes 'crs' or other variables
-        dem = xr.open_dataset(self.dem_filename, engine=self.netcdf_engine, chunks=2*self.netcdf_chunksize)
+        dem = xr.open_dataset(self.dem_filename, engine=self.netcdf_engine, chunks=self.netcdf_chunksize)
         assert 'lat' in dem.coords and 'lon' in dem.coords, 'DEM should be defined as lat,lon grid'
         # define latlon array
         z_array_name = [data_var for data_var in dem.data_vars if len(dem.data_vars[data_var].coords)==2]
@@ -786,7 +786,7 @@ class SBAS:
         
         # open DEM file and find the elevation variable
         # because sometimes grid includes 'crs' or other variables
-        landmask = xr.open_dataset(self.landmask_filename, engine=self.netcdf_engine, chunks=2*self.netcdf_chunksize)
+        landmask = xr.open_dataset(self.landmask_filename, engine=self.netcdf_engine, chunks=self.netcdf_chunksize)
         assert 'lat' in landmask.coords and 'lon' in landmask.coords
         # define latlon array
         z_array_name = [data_var for data_var in landmask.data_vars if len(landmask.data_vars[data_var].coords)==2]
@@ -1044,8 +1044,8 @@ class SBAS:
         intf_ra2ll_file = os.path.join(self.basedir, f'F{subswath}_intf_ra2ll.grd')
         intf_ll2ra_file = os.path.join(self.basedir, f'F{subswath}_intf_ll2ra.grd')
 
-        matrix_ra2ll = xr.open_dataarray(intf_ra2ll_file, engine=self.netcdf_engine, chunks=2*self.netcdf_chunksize)
-        matrix_ll2ra = xr.open_dataarray(intf_ll2ra_file, engine=self.netcdf_engine, chunks=2*self.netcdf_chunksize)
+        matrix_ra2ll = xr.open_dataarray(intf_ra2ll_file, engine=self.netcdf_engine, chunks=self.netcdf_chunksize)
+        matrix_ll2ra = xr.open_dataarray(intf_ll2ra_file, engine=self.netcdf_engine, chunks=self.netcdf_chunksize)
 
         # conversion works for a different 1st grid dimension size
         def ra2ll(grid):
@@ -1101,7 +1101,7 @@ class SBAS:
         lat_min, lat_max = trans[:,4].min(),trans[:,4].max()
 
         # read translation table for the full DEM area
-        trans_ra2ll = xr.open_dataarray(trans_ra2ll_file, engine=self.netcdf_engine, chunks=2*self.netcdf_chunksize)
+        trans_ra2ll = xr.open_dataarray(trans_ra2ll_file, engine=self.netcdf_engine, chunks=self.netcdf_chunksize)
 
         # build ra2ll translation matrix for interferogram coordinates and area only
         # each lan/lon cell has zero or one neighbour radar cell
@@ -1197,7 +1197,7 @@ class SBAS:
             os.remove(intf_ll2ra_file)
 
         # read translation table for the full DEM area
-        trans_ra2ll = xr.open_dataarray(trans_ra2ll_file, engine=self.netcdf_engine, chunks=2*self.netcdf_chunksize)
+        trans_ra2ll = xr.open_dataarray(trans_ra2ll_file, engine=self.netcdf_engine, chunks=self.netcdf_chunksize)
 
         # trans.dat - file generated by llt_grid2rat (r a topo lon lat)"
         trans = self.get_trans_dat(subswath)
@@ -1260,9 +1260,9 @@ class SBAS:
[…]
