### 2022-10-05 f5b5c73
Use Float32 datatype for detrending results as for all other data
