### 2022-10-08 2e8bfea
Change compilation command which produces segfault for make_gaussian_filter on ARM64 Ubuntu 22.04
