### 2022-10-08 f416bdd
Update default NetCDF chunk size from 256x256 to 512x512 for faster lazy calculations on big grid stacks
