### 2022-10-08 4eb5ee8
Always define landmask on the DEM grid
