### 2022-10-08 ddc88dc
Document recent API changes. Fix Ubuntu GMTSAR installation instructions.
