### 2022-10-08 8ae3369
Fix test case to work with the larger default NetCDF chunk size (512)
