+++ tests/S1A_Stack_CPGF_T173.py
@@ -36,8 +36,8 @@ print (sbas.get_topo_ra())
 #Interferograms
 pairs = baseline_pairs[['ref_date', 'rep_date']]
 #decimator = lambda dataarray: dataarray.coarsen({'y': 4, 'x': 4}, boundary='trim').mean()
-# use the DEM resolution SRTM3 (3 sec, 90 m)
-decimator = sbas.pixel_decimator(resolution_meters=90, debug=True)
+# SRTM3 DEM resolution 3 sec, 90 m but use 60 m instead to follow NetCDF chunks size 512x512
+decimator = sbas.pixel_decimator(resolution_meters=60, debug=True)
 sbas.intf_parallel(pairs, wavelength=400, func=decimator)
 # output
 print (pairs)
