### 2022-10-09 0105171
Try to fix weird issue in Docker cross-platform builds
