### 2022-10-09 af1338e
Fix Docker base image and add the Docker howto into README
