+++ docs/README.md
@@ -28,6 +28,20 @@ Satellite interferometry is a complex field of knowledge and usually that's requ
 
 See the project sources and bug tracker on [PyGMTSAR GitHub](https://github.com/mobigroup/gmtsar) and documentation on [PyGMTSAR GitHub Pages](https://mobigroup.github.io/gmtsar/)
 
+### Tutorials: Live Examples in Docker image
+
+Configure your Docker runtime (Preferences -> Resources tab for Docker Desktop) to use 2 CPU cores and 8 GB RAM or 4 CPU cores and 16 GB RAM and so on. Download the Docker image (or build it yourself using the Dockerfile in the repository) and run the container forwarding port 8888 to JupyterLab using this commands inside your command line terminal window:
+
+```
+docker pull mobigroup/pygmtsar
+
+docker run -dp 8888:8888 --name pygmtsar docker.io/mobigroup/pygmtsar
+
+docker logs pygmtsar
+```
+
+See the output for the JupyterLab link and copy and past it into your web browser address line. Also, the donwloaded Docker image can be started in Docker Desktop app - press "RUN" button and define the container name and the port in the opened dialog window (see "Optional settings" for the port number input field) and click on the newly created container to launch it and see the output log with the clickable link.
+
 ## Tutorials: Live Examples on Google Colab
 
 Click on the examples below to run the processing in your own browser without any software installation. That's like to magic and it works.
