### 2022-10-10 d50d7e8
Update a single screenshot for example notebook
