### 2022-10-10 #60 @lwasser
hey @mobigroup i just wanted to check in with you - we are searching for an editor so i'm hopeful that we can get things going soon. i'll stay in touch this week as I continue to search. We are just reestablishing our structures so I truly appreciate your patience here. please ping me with any questions
