### 2022-10-10 #60 @mobigroup
Hi @lwasser thanks for the update. FYI, In the recent days I’ve shared PyGMTSAR Docker image which includes the ready to use environment and the live Jupyter notebooks to easy start locally without Google Colab usage. See below how to use it.

Configure your Docker runtime (Preferences -> Resources tab for Docker Desktop) to use 2 CPU cores and 8 GB RAM or 4 CPU cores and 16 GB RAM and so on. Download the Docker image (or build it yourself using the Dockerfile in the repository) and run the container forwarding port 8888 to JupyterLab using this commands inside your command line terminal window:

```
docker pull mobigroup/pygmtsar​
docker run -dp 8888:8888 --name pygmtsar \
    docker.io/mobigroup/pygmtsar​
docker logs pygmtsar
```
See the output for the JupyterLab link and copy and past it into your web browser address line. Also, the donwloaded Docker image can be started in Docker Desktop app — press “RUN” button and define the container name and the port in the opened dialog window (see “Optional settings” for the port number input field) and click on the newly created container to launch it and see the output log with the clickable link.
