### 2022-10-11 877daae
Try to fix some README formatting issues on PyPI page
