### 2022-10-12 #60 @lwasser
oh that's cool. have you thought about making that image into a binder image? or would it not have enough resources? @mobigroup

and just an update. i think i've found an editor for this. thank you for being so patient. We are just getting the organization structure in place and it just takes time. I think this editor will be back online next monday and then we can kick things off.

In the meantime I have a few additional questions (in the future this will be in the review up front)

1. Do you intend to continue to maintain this package or support this package for the next ~2 years?
2.[ Can you please fill out our survey? We will ask the reviewers for this package to also fill it out. It really helps us improve our organization. ](https://forms.gle/darVrzt6tW27cZ286)

Many thanks in advance for your time and patience here.
