### 2022-10-12 #60 @lwasser
One other note - if you want to work on something while we get organized, you could look into the JOSS side of things. We certaintly don't require this but

1. [you should check that you package is in scope for joss](https://joss.readthedocs.io/en/latest/submitting.html#)
2. [if it is, you can work on your paper.md file - instructions here](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)

The way this works is that once our review is done, JOSS accepts our review and only looks at your paper. You are thus fast tracked through the JOSS review process via our review.
More soon.
