### 2022-10-12 #60 @mobigroup
Hi @lwasser thanks, I'm glad to listen about the progress!

> 1. Do you intend to continue to maintain this package or support this package for the next ~2 years?

Yes, sure.

> 2. Can you please fill out our survey?

Ok, I'll.

> ... paper.md file

Ah, got it, I completely forgot about it!
