### 2022-10-16 #60 @mobigroup
Update: NASA operator [Alaska Satellite Facility](https://www.linkedin.com/company/aksatellite/) has installed GMTSAR on OpenSarLab (OSL). Now you can run PyGMTSAR out of the box right after starting OSL.
https://www.linkedin.com/feed/update/urn:li:activity:6987024568028708864/
