### 2022-10-19 #60 @NickleDave
Hi @mobigroup thank you for the update. It's clear you are doing a lot to make PyGMTSAR accessible for everyone

@lwasser asked me to check in here, in my role as editor in chief.

Just letting you know that we do have a potential editor and reviewers. We're still doing a little more homework about fit and overlap. You should hear back on this issue within a week.

In the meantime, could I just remind you to take this quick (10-15m) survey when you get a chance?
https://docs.google.com/forms/d/e/1FAIpQLSe5cVC-QsYza5adnjVbMDLYha9XSuViT9ykEVl9a6fxjt5eSg/viewform
This helps us understand how well we are serving you and the rest of the Python open science community.
Thanks!
