### 2022-10-19 #60 @lwasser
@mobigroup so sorry i gave david the wrong survey link - [this is the correct one. ](https://forms.gle/eDGHEMUQJWWFahVU7). ive updated David's comment above as well. this is my fault.
