### 2022-10-20 #60 @mobigroup
@NickleDave @lwasser thank you for the information, I’ve filed the survey. Let me note one additional point: there is a big problem for freelancers in science community because we can’t publish a paper or obtain a grant (authors must have an affiliation to an industrial company or a research institute) even for projects completed for well-known industial companies (Google Inc, for example, see some related projects on my GitHub) or research institutes. Surprisingly, arXiv asks for registered experts review but always rejects papers from authors without affiliation too.
