### 2022-10-20 #60 @NickleDave
Thank you @mobigroup, we appreciate you taking the time to do that!

I hear you about how the scientific community may be pushing out freelancers (perhaps unintentionally) because of the requirement for affiliations. I guess this was prompted by a question in the survey? Should we add fields for "freelancer" or "unaffiliated" somewhere?
