### 2022-10-20 #60 @mobigroup
@NickleDave It’d be valuable I think. Maybe my case is unique or maybe not: we can’t be sure without the targeted questionnaire. How about some more options for a respondent occupation? Maybe “freelancer” and “contractor” are not rare cases.
