### 2022-10-20 #60 @NickleDave
That makes sense, I have passed that on to @lwasser.

@mobigroup I just want to update you on where we're at with this.

We really appreciate your enthusiasm and we think there are good reasons to review the package.
But it's still not quite clear to us how `pygmtsar` relates to `gmtsar` and what specific functionality `pygmtsar` provides.

Before we start a review, then, we'd ask you to make changes that achieve the following:
update the readme and docs as necessary to make it clear to a totally naive reader what this package does, how it's structured, and in what way it depends on GMTSAR.

It is of course up to you to decide how best to achieve that given the goals and audience of your package.
But here is a little more detail on how I would do it if I were you, though, if it helps:
Of course it's not your job to teach everyone what fast satellite interferometry is in detail. But there should be clear concrete language that, for example, defines the subject area in ~1-2 sentences, and then proceeds to explain precisely what functionality your code provides that makes it easier to work in this area, e.g. "pygmtsar provides two modules for working with the SBAS and PRM file formats" (as someone who is not a domain expert, I have no idea what these file formats are and they are not defined anywhere in the docstrings or documentation -- to me this indicates that the package can do a lot more to make the science open). Similarly, the README/docs should state up front exactly how pygmtsar and gmtsar are related in terms of code: is pygmtsar it a wrapper with actual c bindings? or does it only assume that GMTSAR is installed and call just one or two functions under the hood? I know that you have a lot of this throughout the docs in one form or another, but it should immediately hit me in the face when I'm reading, I shouldn't have to search for it and piece things together by staring at the source code.

Please let us know if that makes sense. Happy to continue the conversation here to clarify
