### 2022-10-21 #60 @mobigroup
@NickleDave Ah, maybe would you explain on examples? The project documentation below cited from https://mobigroup.github.io/gmtsar/

1. The first lines of the docs explain the relationship between PyGMTSAR and GMTSAR. Please let me know why do you think that’s not enough?

> PyGMTSAR (Python GMTSAR) is an open source project and Python package that provides Sentinel-1 Satellite Interferometry for everyone! While it’s pure Python package under the hood it uses GMTSAR binary tools which should be installed.

> Initially, PyGMTSAR is forked from GMTSAR GitHub repository and lot’s of changes are made to seamlessly call all the binary tools from Python API. For now, all the changes developed for PyGMTSAR project merged into GMTSAR and the both projects use the same binary tools although PyGMTSAR maintains rich Python API for interactive and batch computations and GMTSAR provides a set of shell scripts for the batch processing only. To prevent confusing, below PyGMTSAR means the Python package only and GMTSAR means the binary core tools (while these are available in PyGMTSAR GitHub repository too).

2. The subject area is explained and I actually confused why you’ve found it not informative.

> Why PyGMTSAR?

> PyGMTSAR itself combines powerful Python instrumentary for sophisticated multidementional processing (xarray library) and lazy calculations (dask library) plus parallel computing (dask and joblib libraries) to perform fast and interactive processing on huge datasets. And the best algorithms and numerical computation approaches applied for all the processing steps. There are progressbars and preview plots for the every step and that’s easy to save intermediate results and continue work later on the same or other host. And (thanks to joblib library) that’s safe to interrupt the execution at any time without memory leaks (common for dask-based solutions).

> Thanks to all the powerful Python libraries and the best used algorithms PyGMTSAR is fast and its possible to complete SBAS analysis for 5 years on 800 interferograms in just one day on Apple Air or Apple iMac (8 cores and 16 GB RAM) using 2 TB Sentinel-1 SLC scenes. And PyGMTSAR is user-friendly providing functions to download the required satellite orbit files and DEM and so on. This combination of the human-readable and short code and powerful computing is the key to use PyGMTSAR everywhere from education and to research and more.

3. The functions reference lists the top-level functions with notes is that PyGMTSAR pure Python or GMTSAR binary wrapper. To be honest I don’t know to make it more clear.

> List scenes and orbits (PyGMTSAR original)
…
Framing (combined GMTSAR wrapper and PyGMTSAR original)
…
Build Interferogram (GMTSAR wrapper)
…

Please let me know some exact recommendations how to make the documentation more clear!
