### 2022-10-22 627064b
Move NetCDF parameter to separate class to manage them in the single place
