### 2022-10-22 ec5bee8
Create Dask progress bar based on tqdm progressbar. This progressbar is blocking and so the next command will be executed only when the current execution (in the progressbar) is completed.
