### 2022-10-22 6b4b9c9
Move some functions for grid operations like to gaussian filtering from PRM to datagrid.
