+++ pygmtsar/pygmtsar/datagrid.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
 # Alexey Pechnikov, Oct, 2022, https://github.com/mobigroup/gmtsar
+#from pygmtsar import datagrid
 
 class datagrid:
 
@@ -7,3 +8,154 @@ class datagrid:
     chunksize = 512
     compression = dict(zlib=True, complevel=3, chunksizes=(chunksize, chunksize))
     engine = 'h5netcdf'
+
+    
+    # replacement for GMTSAR gaussians
+    # gauss5x5 = np.genfromtxt('/usr/local/GMTSAR/share/gmtsar/filters/gauss5x5',skip_header=True)
+    # gaussian_kernel(5,1) ~= gauss5x5
+    @staticmethod
+    def gaussian_kernel(size=(5,5), std=(1,1)):
+        """Make 2D Gaussian kernel matrix"""
+        import numpy as np
+        from scipy import signal
+        matrix1 = signal.gaussian(size[0], std=std[0]).reshape(size[0], 1)
+        matrix2 = signal.gaussian(size[1], std=std[1]).reshape(size[1], 1)
+        matrix2d = np.outer(matrix1, matrix2)
+        return matrix2d
+
+#    @staticmethod
+#    def nanconvolve2d(data, kernel, threshold=1/3.):
+#        """
+#        Convolution using generic kernel on a 2D array with NaNs
+#        """
+#        import numpy as np
+#        import scipy.signal
+#        # np.complex128 includes np.float64 real and imagine part
+#        vals = data.astype(np.complex128)
+#        #vals[np.isnan(data)] = 0
+#        #vals[~np.isnan(data)] += 1j
+#        nanmask = np.isnan(data)
+#        vals[nanmask] = 0
+#        vals[~nanmask] += 1j
+#
+#        conv = scipy.signal.convolve2d(vals, 
+#                                      kernel.astype(np.complex128), 
+#                                      mode='same', boundary='symm'
+#                                     )
+#        # suppress incorrect division warning
+#        np.seterr(invalid='ignore')
+#        return np.where(conv.imag >= threshold*np.sum(kernel), np.divide(conv.real, conv.imag), np.nan)
+
+    @staticmethod
+    def nanconvolve2d_gaussian(dataarray, sigmas, truncate):
+        """
+        Lazy convolution using Gaussian kernel on a 2D array with NaNs
+        """
+        import xarray as xr
+        import dask
+        from dask_image.ndfilters import gaussian_filter as dask_gaussian_filter
+
+        da = dask.array.where(dask.array.isnan(dataarray), 0, 1j + dataarray)
+        conv = dask_gaussian_filter(da, sigmas, mode='reflect', truncate=truncate)
+        da_conv = xr.DataArray(conv.real/conv.imag, coords=dataarray.coords, name=dataarray.name)
+        return da_conv
+
+    @staticmethod
+    def nearest_grid(in_grid, search_radius_pixels=300):
+        """
+        Pixel-based Nearest Neighbour interpolation
+        """
+        from scipy.spatial import cKDTree
+        import xarray as xr
+        import numpy as np
+        ys, xs = np.meshgrid(range(in_grid.shape[1]), range(in_grid.shape[0]))
+        ys = ys.reshape(-1)
+        xs = xs.reshape(-1)
+        if isinstance(in_grid, xr.DataArray):
+            zs = in_grid.values.reshape(-1)
+        else:
+            zs = in_grid.reshape(-1)
+        mask = np.where(~np.isnan(zs))
+        # on regular source grid some options should be redefined for better performance
+        tree = cKDTree(np.column_stack((ys[mask],xs[mask])), compact_nodes=False, balanced_tree=False)
[…]
