+++ pygmtsar/pygmtsar/PRM.py
@@ -5,12 +5,6 @@
 #import pytest
 
 class PRM:
-
-    # NetCDF options
-    netcdf_chunksize = 512
-    netcdf_compression = dict(zlib=True, complevel=3, chunksizes=(netcdf_chunksize,netcdf_chunksize))
-    netcdf_engine = 'h5netcdf'
-
     @staticmethod
     def gmtsar_sharedir():
         """
@@ -32,162 +26,6 @@ class PRM:
         if data == '':
             return stderr_data
         return data
-    
-    # replacement for GMTSAR gaussians
-    # gauss5x5 = np.genfromtxt('/usr/local/GMTSAR/share/gmtsar/filters/gauss5x5',skip_header=True)
-    # gaussian_kernel(5,1) ~= gauss5x5
-    @staticmethod
-    def gaussian_kernel(size=(5,5), std=(1,1)):
-        """Make 2D Gaussian kernel matrix"""
-        import numpy as np
-        from scipy import signal
-        matrix1 = signal.gaussian(size[0], std=std[0]).reshape(size[0], 1)
-        matrix2 = signal.gaussian(size[1], std=std[1]).reshape(size[1], 1)
-        matrix2d = np.outer(matrix1, matrix2)
-        return matrix2d
-
-    @staticmethod
-    def nanconvolve2d(data, kernel, threshold=1/3.):
-        """
-        Convolution using generic kernel on a 2D array with NaNs
-        """
-        import numpy as np
-        import scipy.signal
-        # np.complex128 includes np.float64 real and imagine part
-        vals = data.astype(np.complex128)
-        #vals[np.isnan(data)] = 0
-        #vals[~np.isnan(data)] += 1j
-        nanmask = np.isnan(data)
-        vals[nanmask] = 0
-        vals[~nanmask] += 1j
-
-        conv = scipy.signal.convolve2d(vals, 
-                                      kernel.astype(np.complex128), 
-                                      mode='same', boundary='symm'
-                                     )
-        # suppress incorrect division warning
-        np.seterr(invalid='ignore')
-        return np.where(conv.imag >= threshold*np.sum(kernel), np.divide(conv.real, conv.imag), np.nan)
-
-    @staticmethod
-    def nanconvolve2d_gaussian(data, sigmas, truncate):
-        """
-        Convolution using Gaussian kernel on a 2D array with NaNs
-        """
-        import numpy as np
-        from scipy.ndimage import gaussian_filter
-
-        # np.complex128 includes np.float64 real and imagine part
-        vals = data.astype(np.complex128)
-        nanmask = np.isnan(data)
-        vals[nanmask] = 0
-        vals[~nanmask] += 1j
-
-        conv = gaussian_filter(vals, sigma=sigmas, truncate=truncate)
-    
-        # suppress incorrect division warning
-        np.seterr(invalid='ignore')
-        return np.divide(conv.real, conv.imag)
-
-    @staticmethod
-    def nearest_grid(in_grid, search_radius_pixels=300):
-        """
-        Pixel-based Nearest Neighbour interpolation
-        """
-        from scipy.spatial import cKDTree
-        import xarray as xr
-        import numpy as np
-        ys, xs = np.meshgrid(range(in_grid.shape[1]), range(in_grid.shape[0]))
-        ys = ys.reshape(-1)
-        xs = xs.reshape(-1)
-        if isinstance(in_grid, xr.DataArray):
-            zs = in_grid.values.reshape(-1)
-        else:
-            zs = in_grid.reshape(-1)
-        mask = np.where(~np.isnan(zs))
-        # on regular source grid some options should be redefined for better performance
-        tree = cKDTree(np.column_stack((ys[mask],xs[mask])), compact_nodes=False, balanced_tree=False)
[…]
