### 2022-10-22 a7bab15
Wide code refactoring to use lazy computations where it is posssible. Some functions still needs to be updated to work.
