+++ pygmtsar/pygmtsar/SBAS.py
@@ -5,12 +5,6 @@
 #import pytest
 
 class SBAS:
-
-    # NetCDF options
-    netcdf_chunksize = 512
-    netcdf_compression = dict(zlib=True, complevel=3, chunksizes=(netcdf_chunksize,netcdf_chunksize))
-    netcdf_engine = 'h5netcdf'
-    
     import contextlib
     @staticmethod
     @contextlib.contextmanager
@@ -88,8 +82,8 @@ class SBAS:
 
     @staticmethod
     def nearest_grid(in_grid, search_radius_pixels=300):
-        from pygmtsar import PRM
-        return PRM.nearest_grid(in_grid, search_radius_pixels)
+        from pygmtsar import datagrid
+        return datagrid.nearest_grid(in_grid, search_radius_pixels)
 
     def snaphu_config(self, defomax=0, **kwargs):
         return self.PRM().snaphu_config(defomax, **kwargs)
@@ -108,56 +102,43 @@ class SBAS:
         #print ('indexer', indexer)
         return das.loc[indexer]
     
-    def gaussian(self, da, wavelength, truncate=3.0, approximate=False, debug=False):
+    def degaussian(self, dataarray, wavelength, truncate=3.0, resolution_meters=90, debug=False):
         """
-        Gaussian filter for arrays with NaN values.
-            da - input dataarray with NaNs allowed,
+        Lazy Gaussian filter for arrays with NaN values.
+            dataarray - input dataarray with NaNs allowed,
             wavelength - cut-off wavelength [m],
             truncate - filter window size [sigma],
-            approximate - boolean flag to use approximate calculation based on filtering on decimated array,
-            debug - pront debug information.
+            resolution_meters - Gaussian filter processing resolution [m],
+            debug - print debug information.
         Returns filtered dataarray with the same coordinates as input one.
         Fast approximate calculation silently skipped when sigma is less than 64 so the result is always exact for small filters.
         """
         import xarray as xr
         import numpy as np
-        from pygmtsar import PRM
+        from pygmtsar import datagrid
 
+        # input grid can be too large
+        decimator = self.pixel_decimator(resolution_meters=resolution_meters, grid=dataarray, debug=debug)
+        # decimate
+        dataarray_dec = decimator(dataarray)
         # ground pixel size
-        dy, dx = self.pixel_size(da)
+        dy, dx = self.pixel_size(dataarray_dec)
         # gaussian kernel
         sigma_y = np.round(wavelength / dy)
         sigma_x = np.round(wavelength / dx)
-    
         if debug:
-            print ('DEBUG: Gaussian filtering sigma_y, sigma_x', sigma_y, sigma_x)
-    
-        # there is a compromise between accuracy and speed
-        dec = 32
-        # ignore decimation when it is less than or equal to 1
-        dec_y = int(np.floor(sigma_y/dec))
-        dec_x = int(np.floor(sigma_x/dec))
-        
-        if approximate and dec_y > 1 and dec_x > 1:
-            # filter decimated array for faster and less accurate processing
-            if debug:
-                print ('DEBUG: Gaussian filtering decimation dec_y, dec_x', dec_y, dec_x)
-        
-            # decimate and filter
-            da_dec = da.coarsen({'y': dec_y, 'x': dec_x}, boundary='pad').mean()
-            values  = PRM.nanconvolve2d_gaussian(da_dec.values,
-                                                      (sigma_y/dec_y, sigma_x/dec_x),
[…]
