+++ pygmtsar/pygmtsar/__init__.py
@@ -1,2 +1,4 @@
+from .datagrid import datagrid
 from .SBAS import SBAS
 from .PRM import PRM
+from .tqdm_dask import tqdm_dask
