### 2022-10-22 #60 @NickleDave
Hi again @mobigroup, thank you for your detailed reply.
You are correct that you have written something about all of the things I listed, as the text you quoted shows. As I said above
> I know that you have a lot of this throughout the docs in one form or another

Let me explain what I'm looking for in terms of our process and our criteria.
Note that the criteria are here:  https://www.pyopensci.org/contributing-guide/open-source-software-peer-review/intro.html#
And the process is here: https://www.pyopensci.org/contributing-guide/open-source-software-peer-review/aims-and-scope.html

These are the relevant criteria/goals:
1. Limit overlapping functionality in packages
2. Improve usability through documentation
3. Design and plan for long term maintenance

## limit overlapping functionality
This is where we have a strong case to review the package. It's obviously in scope for pyOpenSci. Since you are providing to some extent a wrapper around GMTSAR, and there's no other wrappers, it's likely that there's little overlap in functionality with other packages.

*But* we still need to evaluate this. I can't currently do that because I am not able to find a list of features, unless I take the time to read the GMTSAR docs and learn it myself, then take the time to work with pyGMTSAR to see how it provides that functionality.

There's a couple of ways to address this:
### have a clear list of features in the README.
Notice that I said a *list*, not paragraphs of text. These features should be bullet points with succinct, concrete verbs and nouns so that it's easy for a user to find and understand them.
As shown here, for example: https://www.writethedocs.org/guide/writing/beginners-guide-to-docs/#id1
(I know you are by no means a beginner, even though that link is to a beginners guide. I am trying to provide you with examples from others so you do not feel like I am just arbitrarily making up criteria to be a jerk)
Here's an example of an actual README that does this:
https://github.com/amitmerchant1990/electron-markdownify#key-features
from https://github.com/matiassingers/awesome-readme

Just adding this list alone would go a long way towards helping me and potential reviewers understand the specific functionality provided by pyGMTSAR

### have API documentation
The second way to make the functionality of pyGMTSAR plain as day would be to add API docs, for example like pyGMT provides: https://www.pygmt.org/dev/api/index.html

## 3. Design and plan for long term maintenance
Let me explain this before I say anything about the criterion of documentation.
We consider reviewed packages part of our ecosystem.
https://www.pyopensci.org/contributing-guide/open-source-software-peer-review/intro.html#how-is-pyopensci-different-from-joss-and-other-review-processes
This means that in some sense I am committing that I would be a maintainer of the library if needed. Of course we want you to continue develop the package on your own, but we need to review with the idea in mind that we might be maintaining it or seeking new maintainers.

Currently the way the package is presented gives me the impression that that it would require a very significant investment of time for anyone else to ever begin to understand how to take over development and maintenance. This makes me very hesitant to review.

There's a couple of things that concern me in this respect:
- language in the documentation as if it were a personal project, e.g. use of "I" in the first paragraph of the README, focusing on the developer instead of the package itself.
- I also do not find any kind of developer documentation that explains how someone would build the package. This is particularly concerning giving the complications of a build that involves C and a fork that is simultaneously ahead and behind of the upstream package. I do see that there is CI set up with GitHub Actions--that gives some idea of how to build--but it's not clear to me if it runs on every push or merge.

See also this section in our guidebook about maintenance:
https://www.pyopensci.org/contributing-guide/open-source-software-submissions/author-guide.html#does-your-package-have-a-2-year-maintenance-plan

I am not asking you to commit to two years of maintenance now, but I am saying that I am having a hard time imagining who we could even ask to take over maintenance should you need to step away in the future.

## improve package usability through documentation

Lastly let me talk about the documentation. I hope it will make more sense here in light of the review criteria I have just listed and discussed. Here I am replying directly to the examples you provided.

Please know as I said before that your enthusiasm for the project is quite evident as well as the significant amount of work you have already put in. And there are a lot of great things about the documentation! For example the Google CoLabs notebooks are very close to the kind of vignettes we would like projects to have.

With what I say, I am trying to help you. The main thing I would want to see before we begin a review is a revision of the documentation with an emphasis on concise, clear writing, without run-on sentences and without paragraphs of text except where absolutely needed, e.g. a tutorial. Again, I am not saying this because I enjoy critiquing people's writing; I'm asking for this because it would help me know that you are focused on making the project approachable by as wide an audience as possible, as fits with our criteria, and that we would be able to find someone who could if needed take over maintenance.

> The first lines of the docs explain the relationship between PyGMTSAR and GMTSAR. Please let me know why do you think that’s not enough?

Now that you point me at it again, I do see you state a description of the project first thing in the README. But it's only after I've already read through half a paragraph of text, and it's a very long sentence.

A relevant example of a project description from the pyGMT docs:
https://www.pygmt.org/dev/index.html#about
> PyGMT is a library for processing geospatial and geophysical data and making publication quality maps and figures. It provides a Pythonic interface for the [Generic Mapping Tools (GMT)](https://github.com/GenericMappingTools/gmt), a command-line program widely used in the Earth Sciences.

Notice that this is **only** two sentences and it tells me **exactly** what PyGMT does at the highest level. I do not need any knowledge of GMT to understand.

If I were to rewrite your description of pyGMTSAR, I would boil the second paragraph of your README down to something like the following:
"pyGMTSAR provides easy and fast satellite interferometry (InSAR) processing for Sentinel-1 radar scenes everywhere. Under the hood, pyGMTSAR uses GMTSAR binary command line tools, but all GMTSAR scripts and GMT commands are replaced by Python code using modern and robust algorithms."

> The subject area is explained and I actually confused why you’ve found it not informative.

I am not a domain expert, so what you have written does not tell me a lot. If you really want to reach a wider audience of potential users, contributors, and future maintainers, that do not already have deep knowledge of GMT and GMTSAR, you could additionally provide one more paragraph of just one or two sentences defining what satellite interferometry is, e.g. with a wikipedia link. Here the GMTSAR docs do a good job of this, which of course you are familiar with but I'm just trying to illustrate my point: https://topex.ucsd.edu/gmtsar/
For me the description there is much clearer. It uses the most general terms possible ("InSAR processing system designed for users familiar with [Generic Mapping Tools (GMT)](http://gmt.soest.hawaii.edu/)") -- this gives me a general feel for what GMTSAR does -- instead of specific lingo that only experts will be familiar with ("SBAS analysis for 5 years on 800 interferograms ... using 2 TB Sentinel-1 SLC scenes", I have no idea what that means). In addition, the GMTSAR landing page provides multiple citations that tell someone interested where they could go to read more for background information.

> The functions reference lists the top-level functions with notes is that PyGMTSAR pure Python or GMTSAR binary wrapper.

I am not able to find the text you quoted in either the README or the docs -- can you link me to them?

Are these in the modules themselves? I did see in one or two places you wrote something like "this wraps function gmtscript.sh".

What I want to know is: is there just two classes in pyGMTSAR? I'm not saying it's bad if there are, but I'm saying I can't tell from the docs or code, I would have to install it myself and literally `dir` the imported package. If there's just two classes, how many of the methods directly wrap GMTSAR functions? All of them? Just one or two? If it's just one or two, what do the other methods do, which would be considered extra functionality provided by pyGMTSAR? If pyGMTSAR does provide this extra functionality should be made as explicit as possible: e.g., your list of features and your example code snippets. "pyGMTSAR offers the ability to serialize SBAS files to .json with the `SBAR.jsonify` method" (made up example to illustrate my point)

Other things that would help make the package usage and goals obvious in a clear concise way include:
- Brief demonstration usage (https://www.pyopensci.org/contributing-guide/authoring/overview.html#readme)
  + is there a snippet of just a few lines that demonstrates key functionality of pygmtsar? doesn't need to work by cut-and-paste, just needs to give me an idea of usage + api
- Also providing goals. I know you have this, but again if you want a researcher to use your package, who may not be as familiar as you with the area and is feeling overwhelmed, you should provide this as a list! As in https://www.pygmt.org/latest/#project-goals

Again, I can see that you have put a ton of work into pyGMTSAR. I know I wrote a lot here. All of this feedback is meant to help you reach a wider audience, and to make sure we are set up for a good review experience. Please feel free to follow up with questions and more discussion.
