### 2022-10-23 #60 @mobigroup
Hi @NickleDave your explanation is detailed, thanks! I see I need to reread it some times point-by-point. But I'm afraid you still didn't check the project documentation, right? Please open the documentation to find the answers for some of the questions.

## PyGMTSAR Documentation and the Functions Reference

1. PyGMTSAR documentation is linked on the GitHub https://github.com/mobigroup/gmtsar with GitHub Pages icon:

![Screenshot_2022-10-23_at_11_20_36](https://user-images.githubusercontent.com/7342379/197373597-ae40e0da-20df-4748-9167-9a09673dc234.jpg)

2. PyGMTSAR documentation is on the PyPi package page https://pypi.org/project/pygmtsar/

![Screenshot_2022-10-23_at_11_24_44](https://user-images.githubusercontent.com/7342379/197373695-3f4f742a-4122-4924-a7ef-f9fb0e1d65e5.jpg)

3. PyGMTSAR documentation is linked on the DockerHub page https://hub.docker.com/r/mobigroup/pygmtsar

![Screenshot_2022-10-23_at_11_26_50](https://user-images.githubusercontent.com/7342379/197373727-a645cf56-4fc7-45f3-97be-e5bb5b9096d1.jpg)

And the both PyPI and DockerHub pages available directly from the top of README GitHub page:

![Screenshot_2022-10-23_at_11_28_01](https://user-images.githubusercontent.com/7342379/197373814-119d1227-0ffc-4767-91b9-137d4f38f0a5.jpg)

## PyGMTSAR original and wrapped GMTSAR functions

The functions reference includes the exact note for the original and wrapped PyGMTSR functionality:

![Screenshot_2022-10-23_at_11_32_21](https://user-images.githubusercontent.com/7342379/197373985-9141bc30-82aa-4aaf-94b4-1ec29ae50a24.jpg)

![Screenshot_2022-10-23_at_11_36_19](https://user-images.githubusercontent.com/7342379/197374058-e2e9071f-820e-4245-9698-f3b3cff5a936.jpg)

## Package Installation

There are so many ways to install the package:

1. Google Colab Notebooks install the package and all the dependencies automatically and all the installation instructions available in the notebooks:

2. The package is pre-installed in the Docker image on DockerHub https://hub.docker.com/r/mobigroup/pygmtsar and all the installation instructions available in the Docker file https://github.com/mobigroup/gmtsar/tree/master/docker

The Docker image usage instructions listed on the DockerHub page:

Also, the Docker image build instructions included too.

3. The Python PyGMTSAR library installation for the case when you already have GMTSAR binaries installed is described on the PyPI page https://pypi.org/project/pygmtsar/

![Screenshot_2022-10-23_at_11_44_31](https://user-images.githubusercontent.com/7342379/197374271-ed482f2c-455e-46e5-9342-b097067d1cbe.jpg)

4. Continuous integration scripts (CI on GitHub Actions) provide the installation instructions too, see https://github.com/mobigroup/gmtsar/tree/master/.github/workflows

5. At the end of the day, read the installation instructions in the documentation:

## Brief demonstration usage

The Google Colab Live examples provides the use cases and compare the results to other InSAR systems (that's important for science research purposes) and list the keywords to find the notebook resolving the exact tasks:

Do you need to use landmark and compute a vertical displacement? See the notebook "with landmask applied to interferogram, unwapped phase, and LOS, east-west, vertical displacement results." Do you need to crop your scenes? Look for "cropped subswath processing" notebook. How about atmospheric noise correction? It's easy to find too: "detrending approach to remove atmospheric noise". These are not just code snippets but the ready to use interactive processing pipelines.

## The project goal

That's easy and explained in the project title:

> PyGMTSAR (Python GMTSAR) - Sentinel-1 Satellite Interferometry For Everyone

Even GMTSAR developers were sure that's impossible! You'd find our discussions in GMTSAR GitHub project tracker. InSAR users have 256 - 512 GB RAM and make very complex run scripts for the processing. PyGMTSAR performs the processing even on 16 GB RAM Apple Air. PyGMTSAR provides more functionality and faster processing on Apple Air using interactive Jupyter Notebook than you'd obtain on 512 GB RAM host using custom and complex GMTSAR console scripts. As the result, PyGMTSAR is available even in Google Colab Notebooks but we can't do the same for any other satellite interferometry system as I know. This is the huge difference between any **existing** InSAR system and widely **available** InSAR system PyGMTSAR. And all the technical solutions are targeted to achieve the goal!
