+++ pygmtsar/pygmtsar/datagrid.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python3
 # Alexey Pechnikov, Oct, 2022, https://github.com/mobigroup/gmtsar
-#from pygmtsar import datagrid
 
 class datagrid:
 
@@ -9,6 +8,64 @@ class datagrid:
     compression = dict(zlib=True, complevel=3, chunksizes=(chunksize, chunksize))
     engine = 'h5netcdf'
 
+    @staticmethod
+    def is_ra(grid):
+        dims = grid.dims
+        if 'y' in dims and 'x' in dims:
+            return True
+        return False
+
+    @staticmethod
+    def is_geo(grid):
+        dims = grid.dims
+        if 'lat' in dims and 'lon' in dims:
+            return True
+        return False
+
+    def as_geo(self, da):
+        """
+        Add spatial attributes to allow use rioxarray functions like to .rio.clip([geometry])
+        """
+        import sys
+        assert 'rioxarray' in sys.modules, 'rioxarray module is not found'
+        if self.is_geo(da):
+            epsg = 4326
+            y_dim = 'lat'
+            x_dim = 'lon'
+        else:
+            # fake metrical coordinate system just to perform spatial operations
+            epsg = 3857
+            y_dim = 'y'
+            x_dim = 'x'
+        return da.rio.write_crs(epsg).rio.set_spatial_dims(y_dim=y_dim, x_dim=x_dim)
+
+    @staticmethod
+    def is_same(grid1, grid2):
+        dims1 = grid1.dims
+        dims2 = grid2.dims
+        if 'lat' in dims1 and 'lon' in dims1 and 'lat' in dims2 and 'lon' in dims2:
+            return True
+        if 'y' in dims1 and 'x' in dims1 and 'y' in dims2 and 'x' in dims2:
+            return True
+        return False
+
+    def snaphu_config(self, defomax=0, **kwargs):
+        return self.PRM().snaphu_config(defomax, **kwargs)
+ 
+    @staticmethod
+    def cropna(das):
+        # crop NaNs
+        dims = [dim for dim in das.dims if dim != 'pair' and dim != 'date']
+        dim0 = [dim for dim in das.dims if dim in ['pair', 'date']]
+        assert len(dims) == 2, 'ERROR: the input should be 3D array with "pair" or "date" coordinate'
+        da = das.min(dim0)
+        indexer = {}
+        for dim in dims:
+            da = da.dropna(dim=dim, how='all')
+            dim_min, dim_max = da[dim].min().item(), da[dim].max().item()
+            indexer[dim] = slice(dim_min, dim_max)
+        #print ('indexer', indexer)
+        return das.loc[indexer]
     
     # replacement for GMTSAR gaussians
     # gauss5x5 = np.genfromtxt('/usr/local/GMTSAR/share/gmtsar/filters/gauss5x5',skip_header=True)
@@ -88,74 +145,55 @@ class datagrid:
             return xr.DataArray(values, coords=in_grid.coords, name=in_grid.name)
         return values
 
-    # replacement function for GMT based robust 2D trend coefficient calculations:
-    # gmt trend2d r.xyz -Fxyzmw -N1r -V
-    # gmt trend2d r.xyz -Fxyzmw -N2r -V
-    # gmt trend2d r.xyz -Fxyzmw -N3r -V
-    # https://github.com/GenericMappingTools/gmt/blob/master/src/trend2d.c#L719-L744
-    # 3 model parameters
-    # rank = 3 => nu = size-3
-    @staticmethod
-    def GMT_trend2d(data, rank):
+    #decimator = lambda da: da.coarsen({'y': 2, 'x': 2}, boundary='trim').mean()
[…]
