+++ pygmtsar/pygmtsar/__init__.py
@@ -1,19 +1,8 @@
 # unified progress indicators
 from .tqdm_joblib import tqdm_joblib
 from .tqdm_dask import tqdm_dask
-
-# base operations on NetCDF grid
+# base NetCDF operations and parameters on NetCDF grid
 from .datagrid import datagrid
-
-# GMTSAR wrappers
-from .gmtsar_prm import gmtsar_prm
-from .gmtsar import gmtsar
-# SNAPHU wrapper
-from .snaphu import snaphu
-
-from .orbits import orbits
-from .dem import dem
-from .landmask import landmask
-from .geocode import geocode
+# top level module classes
 from .PRM import PRM
 from .SBAS import SBAS
