### 2022-10-23 df04baf
Enhance classes structure and fix issues
