+++ pygmtsar/pygmtsar/SBAS_sbas.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python3
 # Alexey Pechnikov, Sep, 2021, https://github.com/mobigroup/gmtsar
-from .SBAS_detrend import SBAS_detrend
+from .SBAS_sbas_gmtsar import SBAS_sbas_gmtsar
 
-class SBAS_sbas(SBAS_detrend):
+class SBAS_sbas(SBAS_sbas_gmtsar):
 
     def baseline_table(self):
         import pandas as pd
@@ -281,116 +281,3 @@ class SBAS_sbas(SBAS_detrend):
                     #print (date_a_ref, date_a_rep, date_b_rep)
                     data.append({'A': date_a_ref, 'B': date_a_rep, 'C': date_b_rep})
         return pd.DataFrame(data).sort_values(['A', 'B', 'C'])
-
-    def sbas(self, baseline_pairs, smooth=0, atm=0, debug=False):
-        """
-         USAGE: sbas intf.tab scene.tab N S xdim ydim [-atm ni] [-smooth sf] [-wavelength wl] [-incidence inc] [-range -rng] [-rms] [-dem]
-
-         input:
-          intf.tab             --  list of unwrapped (filtered) interferograms:
-           format:   unwrap.grd  corr.grd  ref_id  rep_id  B_perp
-          scene.tab            --  list of the SAR scenes in chronological order
-           format:   scene_id   number_of_days
-           note:     the number_of_days is relative to a reference date
-          N                    --  number of the interferograms
-          S                    --  number of the SAR scenes
-          xdim and ydim        --  dimension of the interferograms
-          -smooth sf           --  smoothing factors, default=0
-          -atm ni              --  number of iterations for atmospheric correction, default=0(skip atm correction)
-          -wavelength wl       --  wavelength of the radar wave (m) default=0.236
-          -incidence theta     --  incidence angle of the radar wave (degree) default=37
-          -range rng           --  range distance from the radar to the center of the interferogram (m) default=866000
-          -rms                 --  output RMS of the data misfit grids (mm): rms.grd
-          -dem                 --  output DEM error (m): dem.grd
-
-         output:
-          disp_##.grd          --  cumulative displacement time series (mm) grids
-          vel.grd              --  mean velocity (mm/yr) grids
-
-         example:
-          sbas intf.tab scene.tab 88 28 700 1000
-        """
-        import os
-        import subprocess
-        import numpy as np
-        import math
-        import glob
-        import datetime
-
-        # cleanup
-        for filename in glob.glob(os.path.join(self.basedir, 'disp*.grd')):
-            os.remove(filename)
-        filename = os.path.join(self.basedir, 'vel.grd')
-        if os.path.exists(filename):
-            os.remove(filename)
-
-        unwrap = self.open_grids(baseline_pairs[['ref_date', 'rep_date']][:1], 'unwrap')[0]
-        dem = self.get_dem(geoloc=True)
-        prm = self.PRM()
-
-        #N=$(wc -l intf.in   | cut -d ' ' -f1)
-        #S=$(wc -l scene.tab | cut -d ' ' -f1)
-
-        N = len(baseline_pairs)
-        S = len(np.unique(list(baseline_pairs['ref_date']) + list(baseline_pairs['rep_date'])))
-
-        #bounds = self.geoloc().dissolve().envelope.bounds.values[0]
-        llmin, ltmin, llmax, ltmax = self.get_master().dissolve().envelope.bounds.values[0].round(3)
-        lon0 = (llmin + llmax)/2
[…]
