+++ pygmtsar/pygmtsar/SBAS_unwrap_snaphu.py
@@ -1,0 +1,122 @@
+#!/usr/bin/env python3
+# Alexey Pechnikov, Oct, 2022, https://github.com/mobigroup/gmtsar
+from .SBAS_landmask import SBAS_landmask
+
+class SBAS_unwrap_snaphu(SBAS_landmask):
+
+    # -s for SMOOTH mode and -d for DEFO mode when DEFOMAX_CYCLE should be defined in the configuration
+    # DEFO mode (-d) and DEFOMAX_CYCLE=0 is equal to SMOOTH mode (-s)
+    # https://web.stanford.edu/group/radar/softwareandlinks/sw/snaphu/snaphu_man1.html
+    def unwrap(self, subswath, pair, threshold=None, conf=None, func=None, mask=None, conncomp=False,
+               interactive=True, debug=False):
+        import xarray as xr
+        import numpy as np
+        import os
+        import subprocess
+
+        if threshold is None:
+            # set to very low value but still exclude 0 (masked areas)
+            threshold = 1e-6
+
+        # convert user-defined mask to binary mask (NaN values converted to 0)
+        if mask is not None:
+            assert self.is_ra(mask), 'ERROR: mask should be defined in radar coordinates'
+            # crop mask to minimum extent
+            binmask = self.cropna(xr.where(mask>0, 1, np.nan)).fillna(0).astype(bool)
+        else:
+            binmask = 1
+
+        if conf is None:
+            conf = self.PRM(subswath).snaphu_config()
+
+        # extract dates from pair
+        date1, date2 = pair
+
+        basename = os.path.join(self.basedir, f'F{subswath}_{date1}_{date2}_').replace('-','')
+        #print ('basename', basename)
+
+        # input data grids
+        phase_filename = basename + 'phasefilt.grd'
+        corr_filename = basename + 'corr.grd'
+        # output data grids
+        unwrap_filename = basename + 'unwrap.grd'
+        conncomp_filename = basename + 'conncomp.grd'
+
+        # SNAPHU input files
+        phase_in = basename + 'unwrap.phase'
+        corr_in = basename + 'unwrap.corr'
+        # SNAPHU output files
+        unwrap_out = basename + 'unwrap.out'
+        conncomp_out = basename + 'conncomp.out'
+
+        phase = xr.open_dataarray(phase_filename, engine=self.engine, chunks=self.chunksize)
+        corr = xr.open_dataarray(corr_filename, engine=self.engine, chunks=self.chunksize)
+        if mask is not None:
+            phase = phase.reindex_like(binmask)
+            corr = corr.reindex_like(binmask)
+
+        # cleanup from previous runs
+        for tmp_file in [phase_in, corr_in, unwrap_out, conncomp_out,
+                         conncomp_filename, unwrap_filename]:
+            #print ('tmp_file', tmp_file)
+            if os.path.exists(tmp_file):
+                if debug:
+                    print ('DEBUG: remove', tmp_file)
+                os.remove(tmp_file)
+
+        # prepare SNAPHU input files
+        # NaN values are not allowed for SNAPHU phase input file
+        phase.where(~np.isnan(phase),0).astype(np.float32).values.tofile(phase_in)
+        # apply threshold and binary mask
+        corrmasked = corr.where(binmask & (corr>=threshold))
+        # NaN values are not allowed for SNAPHU correlation input file
+        corrmasked.fillna(0).astype(np.float32).values.tofile(corr_in)
+
+        # launch SNAPHU binary (NaNs are not allowed for input but returned in output)
+        argv = ['snaphu', phase_in, str(phase.shape[1]), '-c', corr_in,
+                '-f', '/dev/stdin', '-o', unwrap_out, '-d']
+        if conncomp:
+            argv.append('-g')
+            argv.append(conncomp_out)
[…]
