### 2022-10-23 0363089
Fix issues after the code refactoring
