+++ pygmtsar/pygmtsar/SBAS_landmask.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python3
 # Alexey Pechnikov, Sep, 2021, https://github.com/mobigroup/gmtsar
-from .SBAS_merge import SBAS_merge
+from .SBAS_landmask_gmt import SBAS_landmask_gmt
 
-class SBAS_landmask(SBAS_merge):
+class SBAS_landmask(SBAS_landmask_gmt):
 
     def set_landmask(self, landmask_filename):
         import os
@@ -35,45 +35,3 @@ class SBAS_landmask(SBAS_merge):
         dem = self.get_dem(subswath, geoloc, buffer_degrees)
         return landmask.reindex_like(dem, method='nearest')
 
-    def download_landmask(self, backend='GMT', debug=False):
-        """
-        Use GMT local data or server to download and build landmask on interferogram DEM area
-        """
-        import os
-        import subprocess
-        from tqdm.auto import tqdm
-
-        if self.landmask_filename is not None:
-            print ('NOTE: landmask exists, ignore the command. Use SBAS.set_landmask(None) to allow new landmask downloading')
-            return
-
-        # generate the same as DEM grid
-        landmask_filename = os.path.join(self.basedir, 'landmask.nc')
-
-        dem = self.get_dem()
-        scale = dem.lon.diff('lon')[0].item()
-        llmin = dem.lon.min().item()
-        llmax = dem.lon.max().item()
-        ltmin = dem.lat.min().item()
-        ltmax = dem.lat.max().item()
-
-        # helper function to run external commands
-        def run_cmd(argv):
-            if debug:
-                print ('DEBUG: argv', argv)
-            cmd = subprocess.run(argv, capture_output=True)
-            if cmd.returncode != 0:
-                print (cmd.stderr)
-                raise Exception('DEM processing error using GMT backend')
-            if debug:
-                print (cmd.stderr)
-                print (cmd.stdout)
-
-        # use GMT commands pipeline to download and preprocess the DEM
-        with tqdm(desc='Landmask Downloading', total=1) as pbar:
-            argv = ['gmt', 'grdlandmask', f'-G{landmask_filename}',
-                    f'-R{llmin}/{llmax}/{ltmin}/{ltmax}', f'-I{scale}', '-V', '-NNaN/1', '-Df']
-            run_cmd(argv)
-            pbar.update(1)
-
-        self.landmask_filename = landmask_filename

[…]
