### 2022-10-23 41ef757
Simplify code
