### 2022-10-23 2d2aabb
Fix backward-compatibility for detrend_parallel() function
