### 2022-10-23 15ac00f
Fix backward-compatibility for detrend_parallel() function
