### 2022-10-23 f5efdda
Update documentation following recent changes in detrending function
