### 2022-10-24 #60 @NickleDave
Hi @mobigroup, yes let's go point-by-point. I'm sorry for the lengthy response that was probably overwhelming.

Thanks for the screenshots that show the API documentation.

Can you please share a link to that? I did look at your docs, many times, but obviously I was not able to find it.
