### 2022-10-24 #60 @mobigroup
@NickleDave  The documentation https://mobigroup.github.io/gmtsar/ includes "Reference" chapter right after "Tutorials: Live Examples on Google Colab" chapter:
