### 2022-10-26 de80571
Fix hard-coded parameter
