+++ pygmtsar/pygmtsar/SBAS_geocode.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 # Alexey Pechnikov, Sep, 2021, https://github.com/mobigroup/gmtsar
 from .SBAS_sbas import SBAS_sbas
+from .tqdm_dask import tqdm_dask
 
 class SBAS_geocode(SBAS_sbas):
 
@@ -119,162 +120,47 @@ class SBAS_geocode(SBAS_sbas):
             os.remove(intf_ra2ll_file)
         intf_ra2ll.to_netcdf(intf_ra2ll_file, encoding={'intf_ra2ll': self.compression}, engine=self.engine)
 
-    def ra2ll(self, subswath, debug=False):
-        """
-        Create radar to geographic coordinate transformation matrix for DEM grid using geocoding table trans.dat
-        """
-        from scipy.spatial import cKDTree
-        import xarray as xr
-        import numpy as np
-        import os
-
-        trans_ra2ll_file = os.path.join(self.basedir, f'F{subswath}_trans_ra2ll.grd')
-
-        if os.path.exists(trans_ra2ll_file):
-            os.remove(trans_ra2ll_file)
-
-        # trans.dat - file generated by llt_grid2rat (r a topo lon lat)"
-        trans = self.get_trans_dat(subswath)
-        lon_min, lon_max = trans[:,3].min(),trans[:,3].max()
-        lat_min, lat_max = trans[:,4].min(),trans[:,4].max()
-
-        #dem = xr.open_dataset(in_dem_gridfile)
-        #dem = self.get_dem(geoloc=True)
-        dem = self.get_dem(geoloc=True).sel(lat=slice(lat_min, lat_max), lon=slice(lon_min, lon_max))
-
-        trans_latlons = np.stack([trans[:,4],trans[:,3]], axis=1)
-        dem_lats, dem_lons = xr.broadcast(dem.lat,dem.lon)
-        dem_latlons = np.stack([dem_lats.values.reshape(-1),dem_lons.values.reshape(-1)], axis=1)
-
-        tree = cKDTree(trans_latlons, compact_nodes=False, balanced_tree=False)
-        # use accurate distance limit as a half of the cell diagonal
-        dlat = dem.lat.diff('lat')[0]
-        dlon = dem.lon.diff('lon')[0]
-        distance_limit = np.sqrt((dlat/2.)**2 + (dlon/2.)**2) + 1e-6
-        d, inds = tree.query(dem_latlons, k = 1, distance_upper_bound=distance_limit, workers=8)
-
-        # produce the same output array as dataset to be able to add global attributes
-        trans_ra2ll = xr.zeros_like(dem).rename('trans_ra2ll')
-        trans_ra2ll.values = np.where(~np.isinf(d), inds, -1).reshape(dem.shape)
-        # magic: add GMT attribute to prevent coordinates shift for 1/2 pixel
-        #trans_ra2ll.attrs['node_offset'] = 1
-        # save to NetCDF file
-        trans_ra2ll.to_netcdf(trans_ra2ll_file, encoding={'trans_ra2ll': self.compression}, engine=self.engine)
-
-    # a single-step translation
-    # see also two-step translation ra2ll & intf_ra2ll_matrix
-    def intf_ll2ra_matrix(self, subswath, intf_grids, debug=False):
-        """
-        Create geographic to radar coordinate transformation matrix for DEM grid (F?_trans_ra2ll.grd)
-        """
-        from scipy.spatial import cKDTree
-        import xarray as xr
-        import numpy as np
-        import os
-
-        # use 2D grid grom the pairs stack
-        # sometimes interferogram grids are different for one azimuth line so use the largest grid
[…]
