### 2022-10-26 #60 @NickleDave
Thank you @mobigroup -- I do see the reference documentation now if I scroll past the tutorials.

This helps me see that many (most?) of the methods on the `SBAS` class are denoted as unique to `pygmtsar`.
