### 2022-10-27 0b7c562
Enhance ll2ra() inverse geocoding function to be lazy
