+++ pygmtsar/pygmtsar/SBAS_incidence.py
@@ -4,25 +4,15 @@ from .SBAS_geocode import SBAS_geocode
 
 class SBAS_incidence(SBAS_geocode):
 
-    def get_sat_look(self, subswath=None):
+    def get_sat_look(self):
         import xarray as xr
         import os
 
-        if subswath is None:
-            # process all the subswaths
-            subswaths = self.get_subswaths()
-        else:
-            # process only one subswath
-            subswaths = [subswath]
+        sat_look_file = self.get_filenames(None, None, 'sat_look')
+        assert os.path.exists(sat_look_file), 'ERROR: satellite looks grid missed. Build it first using SBAS.sat_look_parallel()'
+        sat_look = xr.open_dataset(sat_look_file, engine=self.engine, chunks=self.chunksize)
 
-        sat_looks = []
-        for subswath in subswaths:
-            sat_look_file = os.path.join(self.basedir, f'F{subswath}_sat_look.grd')
-            assert os.path.exists(sat_look_file), 'ERROR: satellite looks grid missed. Build it first using SBAS.sat_look_parallel()'
-            sat_look = xr.open_dataset(sat_look_file, engine=self.engine, chunks=self.chunksize)
-            sat_looks.append(sat_look)
-
-        return sat_looks[0] if len(sat_looks)==1 else sat_looks
+        return sat_look
 
     #gmt grdmath unwrap_mask.grd $wavel MUL -79.58 MUL = los.grd
     def los_displacement_mm(self, unwraps):
@@ -32,12 +22,11 @@ class SBAS_incidence(SBAS_geocode):
         los_disp = scale*unwraps
         return los_disp
 
-    def incidence_angle(self, subswath=None):
+    def incidence_angle(self):
         import xarray as xr
         import numpy as np
 
-        subswath = self.get_subswath(subswath)
-        sat_look = self.get_sat_look(subswath)
+        sat_look = self.get_sat_look()
         incidence_ll = np.arctan2(np.sqrt(sat_look.look_E**2 + sat_look.look_N**2), sat_look.look_U).rename('incidence_angle')
         return incidence_ll
 
