+++ pygmtsar/pygmtsar/SBAS_geocode.py
@@ -5,202 +5,123 @@ from .tqdm_dask import tqdm_dask
 
 class SBAS_geocode(SBAS_sbas):
 
-    def intf_ra2ll(self, subswath=None, grids=None, debug=False):
-        """
-        Geocoding function based on interferogram geocode matrix to call from open_grids(geocode=True)
-        """
-        from tqdm.auto import tqdm
-        import joblib
+    def geocode_parallel(self, pairs):
+        # build trans_dat and topo_ra grids
+        self.topo_ra_parallel()
+        # transforms for interferogram grid
+        intf_grid = self.open_grids(pairs, 'phasefilt').min('pair')
+        # build geographic coordinates transformation matrix for landmask and other grids
+        self.intf_ll2ra_matrix_parallel(intf_grid)
+        # build radar coordinates transformation matrix for the interferograms grid stack        
+        self.intf_ra2ll_matrix_parallel(intf_grid)
+    
+##########################################################################################
+# TODO ra2ll
+##########################################################################################
+    def intf_ra2ll_matrix(self, intf_grid, interactive=False):
         import xarray as xr
         import numpy as np
         import os
-
-        # that's possible to miss the first argument subswath
-        assert subswath is not None or grids is not None, 'ERROR: define input grids'
-        if grids is None:
-            grids = subswath
-            subswath = None
-
-        # helper check
-        if not 'y' in grids.dims and 'x' in grids.dims:
-            print ('NOTE: the grid is not in radar coordinates, miss geocoding')
-            return grids
-
-        # check if subswath exists or return a single subswath for None
-        subswath = self.get_subswath(subswath)
-
-        intf_ra2ll_file = os.path.join(self.basedir, f'F{subswath}_intf_ra2ll.grd')
-        intf_ll2ra_file = os.path.join(self.basedir, f'F{subswath}_intf_ll2ra.grd')
-
-        matrix_ra2ll = xr.open_dataarray(intf_ra2ll_file, engine=self.engine, chunks=self.chunksize)
-        matrix_ll2ra = xr.open_dataarray(intf_ll2ra_file, engine=self.engine, chunks=self.chunksize)
-
-        # conversion works for a different 1st grid dimension size
-        def ra2ll(grid):
-            # input and transform grids should be the same
-            grid = grid.reindex_like(matrix_ll2ra)
-            # some interferograms have different y dimension and matrix has the largest
-            # crop matrix y dimension when it is larger than current interferogram
-            matrix_ra2ll_valid = xr.where(matrix_ra2ll<grid.size, matrix_ra2ll, -1)
-            da_ll = xr.DataArray(np.where(matrix_ra2ll>=0, grid.values.reshape(-1)[matrix_ra2ll_valid], np.nan),
-                coords=matrix_ra2ll_valid.coords)
-            return da_ll
-
-    #        def ra2ll(grid):
-    #            return xr.DataArray(np.where(matrix_ra2ll>=0, grid.values.reshape(-1)[matrix_ra2ll], np.nan),
-    #                coords=matrix_ra2ll.coords)
-
-        # process single 2D raster
-        if len(grids.dims) == 2:
-            return ra2ll(grids)
-
-        # process a set of 2D rasters
-        with self.tqdm_joblib(tqdm(desc='Geocoding', total=len(grids))) as progress_bar:
-            grids_ll = joblib.Parallel(n_jobs=-1)(joblib.delayed(ra2ll)(grids[item]) for item in range(len(grids)))
-        grids_ll = xr.concat(grids_ll, dim=grids.dims[0])
-
[…]
