### 2022-10-27 #60 @NickleDave
Hi again @mobigroup.  Thank you for your continued patience.

As I said above, it seems pretty clear that the functionality of `pygmtsar` is within scope for pyOpenSci. And with your help, I was able to understand that `pygmtsar` provides many features in addition the functionality of `GMTSAR` that it wraps. It's also clear you are putting a lot of work into developing and promoting the package, from the work you are putting in to the README now, and the questions you are handling from users in the issues.

But I don't think we can review at this time, unless you would be able to:
- [ ] greatly simply the development workflow
  - [ ] so that `pygmtsar` is clearly its own package, not a fork of `GMTSAR`
  - [ ] and so that you are not relying on a branch that is simultaneously behind and ahead the main `GMTSAR` branch
- [ ] **and** *thoroughly* document that workflow, with a separate section of the documentation for developers, to make it as easy as possible for someone else to contribute and if needed to help with maintenance. For example like the dev docs provided by [numpy](https://numpy.org/doc/stable/dev/index.html#devindex), [pandas](https://pandas.pydata.org/docs/development/index.html#development), [scikit-learn](https://scikit-learn.org/dev/developers/index.html), and [pygmt](https://www.pygmt.org/dev/contributing.html#contributing-code).

As I stated in the comment above, this is because by reviewing and accepting a package into the pyOpenSci ecosystem, we are committing to [support maintenance of the package](https://www.pyopensci.org/contributing-guide/software-peer-review-guide/author-guide.html#do-you-plan-to-continue-to-maintain-your-package). That includes finding another maintainer should you need to step away. It is unclear to us whether we would be able to guarantee that for `pygmtsar` if we went ahead with the review.

I am sorry these criteria were not more clearly documented in our guidebook. We have not run into this exact situation before so we are figuring out how to handle it on the fly. We will learn from this review process and provide clearer standards, in particular for packages that are wrappers around lower-level languages.

If you feel like you cannot make these changes to the development workflow right now, please know I very much understand as a developer myself. We would still be open to reviewing in the future.

You might consider submitting for review with [JOSS](https://joss.readthedocs.io/en/latest/index.html) in the meantime, which provides a similar review process. That way you could still get credit and get feedback from reviewers without needing to adapt to our specific standards.

Please let me know if the reasoning I am giving for not going forward with the review is clear.
