### 2022-10-28 #60 @mobigroup
@NickleDave I'm sorry but it looks inappropriate. I spent a lot of time answering on your questions (and attaching screenshots because you didn't read the project documentation and wrote again and again about missed documentation) and today you rejecting all the work referencing to the 3 dummy points:

1. "so that pygmtsar is clearly its own package, not a fork of GMTSAR" - Anyone can check the GitHub repository for all the changes in PyGMTSAR files to be sure in the authorship.

2. "and so that you are not relying on a branch that is simultaneously behind and ahead the main GMTSAR branch". - why do you ask me about the different project development? While I work on GMTSAR software too but that's obviously is not related to your review. Even more, the repository is synced to upstream GMTSAR because I validate and add the upstream changes: https://github.com/mobigroup/gmtsar/commit/09f3723e6f9f58d06e8cc3c913a4d2a647dd0939

3. "For example like the dev docs provided by [numpy], [pandas], [scikit-learn], and [pygmt]."

You are referencing to adopted GitHub documentation for developers available on https://docs.github.com/en/get-started/quickstart/contributing-to-projects How does it related to your review?

@lwasser Please let me know is it your official position to insult authors ("unless you would be able to: so that pygmtsar is clearly its own package, not a fork of GMTSAR") to reject an application?  Obviously, I did'n stole the ownership and it's very easy to prove using GitHub commits record and @NickleDave have no any right to blame me or anyone else!
