### 2022-10-28 #60 @NickleDave
Hi @mobigroup

> I spent a lot of time answering on your questions

I know this and I completely understand your frustration and impatience.

> attaching screenshots

You are right that I could have scrolled through the entire single page of documentation the first few times I read it, so that I could find the API reference at the bottom.
I am happy to provide more feedback about ways you could structure your documentation so that it is immediately obvious to a new user how to find what they are looking for. I have omitted that feedback from my last comment, because I want to focus on the main reason we have chosen not to proceed with a review at this time.

> "unless you would be able to: so that pygmtsar is clearly its own package, not a fork of GMTSAR"

I think my intent was not clear here; by no means do I mean to say that you are stealing ownership.
It's just not very clear to a new user who visits the repo that `pygmtsar` is a separate package, because it's a fork, and still has the name `GMTSAR`. My goal is to help you make it obvious to a new user that they are looking at a library called `pygmtsar`, not a fork of `GMTSAR`. That is a minor point that is also not related to the main reason we decided not to review. Again I apologize if my language sounded accusatory, that was not my goal.

> You are referencing to adopted GitHub documentation

I am sorry if my reasons for sharing these links was not clear. I am not asking you to provide an entire section on how to use GitHub.

I am asking you to provide a single page with a clear verbal description of how you develop and build `pygmtsar`. This would explain the logic of the build scripts, and any other information unique to your development workflow, such as how you validate and add upstream changes.

If you can provide that single page clearly documenting how you develop, then it will help us understand what would be required to maintain the project.

Does that make sense?
