### 2022-10-28 #60 @mobigroup
@NickleDave Thanks, that sounds better. At the same time:

1. The Python package for review is https://pypi.org/project/pygmtsar/ and it includes PyGMTSAR Python codes only.
The Python sources placed in the sonamed GitHub project directory of my fork of GMTSAR repository because these sources are required for the development to provide the projects compatibility (PyGMTSAR can reproduce GMTSAR results for research purposes). GMTSAR package itself includes 3rd party SNAPHU software and it doesn’t prevent GMTSAR to be a well known science package. And only you see some issue when 3rd party open source software with the license and copyright included into GitHub repository.

2. The Python package includes common setup.py which can be used by any Python package manager to install it. That’s unified and not unique. The Python package has common GitHub continuous integration (CI) scripts named GitHub Actions which running to test every commit or pull request. And it’s unified and not unique too. The package provides Dockerfile to build the Docker images by common way. And again that’s common not unique tool. At the end of the day PyGMTSAR has the unique functions in standard Python package in standard GitHub repository with standard continuous integration for standard git commits and pull requests for all developers. Actually, PyGMTSAR follows the best practices for GitHub repositories.

3. What is the real reason to reject the review process? When you suddenly send weird new requirement to have a unique development process instead of unique functions I suppose the reason is not technical and not related to the project. Please explain why PyGMTSAR can’t be reviewed when it’s satisfying for all the public requirements? Why only for me you invented the new requirement to have a unique development process (you didn’t ask me if the development process has any unique points which need to be documented but you strongly require to have the unique development process)? Maybe just to find any reason for the rejection?
