### 2022-10-28 #60 @NickleDave
@mobigroup

We are not rejecting a review. I should have phrased that differently in my last reply.

As I said in the reply before that, if you can provide documentation of the development process, that will help us understand what is required. If we feel that we would be able to find new maintainers given the requirement, then we could proceed with a review.

You have just helped me better understand development. Thank you.

> Why only for me you invented the new requirement

Like I said, we have not dealt with this situation before and I apologize that we are working out how to handle it now. I get why you feel like I am making up some new criterion. I acknowledge that we do not currently have very specific language in our guidebook about what kind of documentation a project needs to have about its development. I have raised this issue to add it to avoid this problem in the future: https://github.com/pyOpenSci/python-package-guide/issues/8

Please give me and the rest of the editors until Monday at the latest to carefully read your reply and understand what you have explained. I want to make sure I do not make this process any more frustrating for you because I have not thoroughly understood what you are telling me here. I will reply in more detail as soon as I have a chance.
