### 2022-10-28 74308b9
Add linear index to trans_dat table to prevent error to calculate it on the fly later
