### 2022-10-29 6cb5c8d
Enhance lazy processing
