+++ pygmtsar/pygmtsar/SBAS_geocode.py
@@ -16,7 +16,7 @@ class SBAS_geocode(SBAS_sbas):
         self.intf_ra2ll_matrix_parallel(intf_grid)
     
 ##########################################################################################
-# TODO ra2ll
+# ra2ll
 ##########################################################################################
     def intf_ra2ll_matrix(self, intf_grid, interactive=False):
         import xarray as xr
@@ -115,9 +115,8 @@ class SBAS_geocode(SBAS_sbas):
         return xr.DataArray((grids_ll[0] if len(grids.dims)==2 else grids_ll), coords=matrix_ra2ll.coords)
 
 ##########################################################################################
-# TODO ll2ra
+# ll2ra
 ##########################################################################################
-
     def intf_ll2ra_matrix(self, intf_grid, interactive=False):
         from scipy.spatial import cKDTree
         import dask
@@ -128,17 +127,19 @@ class SBAS_geocode(SBAS_sbas):
         # trans.dat - file generated by llt_grid2rat (r a topo lon lat)"
         trans_dat = self.get_trans_dat()
         trans_blocks_extents = self.get_trans_dat_blocks_extents(None)
-        trans_idx = dask.array.arange(trans_dat.ll.size, dtype=np.uint32)\
-                    .reshape(trans_dat.ll.shape)*xr.ones_like(trans_dat.ll, dtype=np.uint32)
-        azis = xr.DataArray(intf_grid.y, dims=['y'], coords={'y': intf_grid.y}).chunk(self.chunksize)
-        rngs = xr.DataArray(intf_grid.x, dims=['x'], coords={'x': intf_grid.x}).chunk(self.chunksize)
+
+        # target grid - do not use coordinate names y,x, because the output grid saved as (y,y) in this case...
+        azis = xr.DataArray(intf_grid.y.values, dims=['a'], coords={'a': intf_grid.y.values}).chunk(-1)
+        rngs = xr.DataArray(intf_grid.x.values, dims=['r'], coords={'r': intf_grid.x.values}).chunk(-1)
         azis, rngs = xr.broadcast(azis, rngs)
-        _, azis, rngs = xr.unify_chunks(intf_grid, azis, rngs)
+        # unify chunks to have the same dask blocks - pay attention dimensions are renamed
+        _, azis, rngs = xr.unify_chunks(intf_grid.rename({'y': 'a', 'x': 'r'}), azis, rngs)
+        assert intf_grid.chunks == azis.chunks and intf_grid.chunks == rngs.chunks, \
+                'Chunks are not equal'
 
         def calc_intf_ll2ra_matrix(azi, rng):
-            # check thr arguments
+            # check arguments
             assert azi.shape == rng.shape, f'ERROR: {azi.shape} != {rng.shape}'
-
             # check the selected area bounds
             ymin, ymax, xmin, xmax = azi.min(), azi.max(), rng.min(), rng.max()
             #print ('ymin, ymax', ymin, ymax, 'xmin, xmax', xmin, xmax)
@@ -146,20 +147,20 @@ class SBAS_geocode(SBAS_sbas):
             ymask = (trans_blocks_extents[:,3]>=ymin-1)&(trans_blocks_extents[:,2]<=ymax+1)
             xmask = (trans_blocks_extents[:,5]>=xmin-1)&(trans_blocks_extents[:,4]<=xmax+1)
             blocks = trans_blocks_extents[ymask&xmask]
-            print ('trans_dat blocks', blocks.astype(int))
+            #print ('trans_dat blocks', blocks.astype(int))
 
             blocks_azis = []
             blocks_rngs = []
             blocks_idxs = []
             for iy, ix in blocks[:,:2].astype(int):
                 #print ('iy, ix', iy, ix)
[…]
