### 2022-10-29 084ec51
Start sbas_parallel() function code refactoring
