### 2022-10-29 0193aab
Fix Xarray related issue with coordinates. Add optimzation to cropt valid area only for trans_dat calculation.
