### 2022-10-29 6f6dee6
Use trans_dat area instead of DEM for landmask processing
