### 2022-10-29 262af10
Drop default subswath value from function call
