+++ pygmtsar/pygmtsar/SBAS_geocode.py
@@ -126,7 +126,7 @@ class SBAS_geocode(SBAS_sbas):
 
         # trans.dat - file generated by llt_grid2rat (r a topo lon lat)"
         trans_dat = self.get_trans_dat()
-        trans_blocks_extents = self.get_trans_dat_blocks_extents(None)
+        trans_blocks_extents = self.get_trans_dat_blocks_extents()
 
         # target grid - do not use coordinate names y,x, because the output grid saved as (y,y) in this case...
         azis = xr.DataArray(intf_grid.y.values, dims=['a'], coords={'a': intf_grid.y.values}).chunk(-1)
