### 2022-10-29 0c9561a
Enhance work area selection
