+++ pygmtsar/pygmtsar/SBAS_trans.py
@@ -65,6 +65,9 @@ class SBAS_trans(SBAS_stack):
                 .astype(np.float32).reshape(z.shape[0], z.shape[1], 5)
             return coords_ra
 
+        ################################################################################
+        # define valid area checking every 10th pixel per the both dimensions 
+        ################################################################################
         # do not use coordinate names lat,lon because the output grid saved as (lon,lon) in this case...
         dem = self.get_dem(geoloc=True).rename({'lat': 'yy', 'lon': 'xx'})
         # prepare lazy coordinate grids
@@ -74,14 +77,6 @@ class SBAS_trans(SBAS_stack):
         # unify chunks
         lats = lats.chunk(dem.chunks)
         lons = lons.chunk(dem.chunks)
-        # calculate linear index
-        lat_idx = xr.DataArray(np.arange(lat.size, dtype=np.uint32), dims=['yy'], coords=lat.coords).chunk(-1)
-        lon_idx = xr.DataArray(np.arange(lon.size, dtype=np.uint32), dims=['xx'], coords=lon.coords).chunk(-1)
-        lat_idxs, lon_idxs = xr.broadcast(lat_idx, lon_idx)
-        # unify chunks
-        idxs = (lat_idxs*lon_idx.size + lon_idxs).chunk(dem.chunks)
-        #print ('idxs', idxs)
-        assert dem.size - idxs.max() == 1, 'Linear index incorrect'
 
         # xarray wrapper for fast valid area lookup
         raell = xr.apply_ufunc(
@@ -96,26 +91,37 @@ class SBAS_trans(SBAS_stack):
             dask_gufunc_kwargs={'output_sizes': {'raell': 5}},
             kwargs={'subswath': subswath, 'binary': True}
         ).data.reshape(-1,5)
-    
+
         # select valid azimuths only
-        yvalid, num_patch = self.PRM(subswath).get('num_valid_az', 'num_patches')
+        rng_max, yvalid, num_patch = self.PRM(subswath).get('num_rng_bins', 'num_valid_az', 'num_patches')
         azi_max = yvalid * num_patch
         #print ('azi_max', azi_max)
-        raell = raell[(raell[...,1]>=0)&(raell[...,1]<=azi_max)]
-    
+        raell = raell[(raell[...,1]>=0)&(raell[...,1]<=azi_max)&(raell[...,0]>=0)&(raell[...,0]<=rng_max)]
+
         lat_min, lat_max = dask.compute(raell[...,-1].min(), raell[...,-1].max())
         #print ('lat_min, lat_max', lat_min, lat_max)
         lon_min, lon_max = dask.compute(raell[...,-2].min(), raell[...,-2].max())
         #print ('lon_min, lon_max', lon_min, lon_max)
-        azi_min, azi_max = dask.compute(raell[...,1].min(), raell[...,1].max())
+        #azi_min, azi_max = dask.compute(raell[...,1].min(), raell[...,1].max())
         #print ('azi_min, azi_max', azi_min, azi_max)
-    
-        # crop the valid area only
+
+        ################################################################################
+        # process the valid area
+        ################################################################################
+        # crop valid area only
         dem  = dem.sel( yy=slice(lat_min, lat_max), xx=slice(lon_min, lon_max))
         lats = lats.sel(yy=slice(lat_min, lat_max), xx=slice(lon_min, lon_max))
         lons = lons.sel(yy=slice(lat_min, lat_max), xx=slice(lon_min, lon_max))
-        idxs = idxs.sel(yy=slice(lat_min, lat_max), xx=slice(lon_min, lon_max))
-    
+
+        # calculate linear index
[…]
