### 2022-10-29 7a308bd
Add sat_look_parallel() function
