+++ pygmtsar/pygmtsar/SBAS_incidence.py
@@ -48,3 +48,65 @@ class SBAS_incidence(SBAS_geocode):
         los_disp = self.los_displacement_mm(unwraps)
         incidence_ll = self.incidence_angle()
         return sign * los_disp/np.sin(incidence_ll)
+
+    def sat_look_parallel(self, n_jobs=-1, interactive=False):
+        import numpy as np
+        import xarray as xr
+        from tqdm.auto import tqdm
+        import joblib
+        import os
+
+        def SAT_look(ilat, ilon):
+            # lazy dask arrays
+            lats, lons = xr.broadcast(coordlat[ilat], coordlon[ilon])
+            data = grid_ll.sel(lat=xr.DataArray(lats.values.ravel()),
+                                 lon=xr.DataArray(lons.values.ravel()),
+                                 method='nearest').compute()
+            coords = np.column_stack([lons.values.ravel(), lats.values.ravel(), data.values.ravel()])
+            # look_E look_N look_U
+            look = self.PRM().SAT_look(coords, binary=True)\
+                                     .reshape(-1,6)[:,3:].astype(np.float32)
+            # prepare output as xarray dataset
+            dims = ['lat', 'lon']
+            coords = coords={'lat': coordlat[ilat], 'lon':coordlon[ilon]}
+            look_E = xr.DataArray(look[:,0].reshape(lats.shape), dims=dims, coords=coords, name='look_E')
+            look_N = xr.DataArray(look[:,1].reshape(lats.shape), dims=dims, coords=coords, name='look_N')
+            look_U = xr.DataArray(look[:,2].reshape(lats.shape), dims=dims, coords=coords, name='look_U')
+            return xr.merge([look_E, look_N, look_U])
+
+        # take elevation values on the interferogram area in geographic coordinates
+        grid_ll = self.get_intf_ra2ll()
+        dem = self.get_dem()
+        grid_ll = dem.interp_like(grid_ll)
+    
+        lats, lons = grid_ll.data.numblocks
+        latchunks, lonchunks = grid_ll.chunks
+        coordlat = np.array_split(grid_ll.lat, np.cumsum(latchunks))
+        coordlon = np.array_split(grid_ll.lon, np.cumsum(lonchunks))
+        with self.tqdm_joblib(tqdm(desc=f'SAT_look Computing', total=lats*lons)) as progress_bar:
+            sat_look = joblib.Parallel(n_jobs=n_jobs)(joblib.delayed(SAT_look)(ilat, ilon) \
+                                           for ilat in range(lats) for ilon in range(lons))
+        # concatenate the chunks
+        if xr.__version__ == '0.19.0':
+            # for Google Colab
+            sat_look = xr.merge(sat_look)
+        else:
+            # for modern xarray versions
+            sat_look = xr.combine_by_coords(sat_look)
+
+        # fill NODATA
+        sat_look = xr.where(grid_ll != self.noindex, sat_look, np.nan)
+    
+        if interactive:
+            return sat_look
+
+        # magic: add GMT attribute to prevent coordinates shift for 1/2 pixel
+        #ds.attrs['node_offset'] = 1
+        # save to NetCDF file
+        sat_look_file = self.get_filenames(None,None,'sat_look')
+        # cleanup before creating the new file
+        if os.path.exists(sat_look_file):
+            os.remove(sat_look_file)
+        sat_look.to_netcdf(sat_look_file,
+                     encoding={var:self.compression for var in sat_look.data_vars},
+                     engine=self.engine)
