### 2022-10-30 c34f7b4
Add function find_grids()
