### 2022-10-31 #60 @NickleDave
Hi @mobigroup thank you for your reply, and for your patience with this review.
I think I see now your development workflow.

To update you: we are continuing to talk with maintainers that develop similar packages so we can better understand what would be required for maintenance.

We should have concrete feedback for you by the end of the week. Again I am sorry it is not sooner -- @lwasser is dealing with a family situation right now. We appreciate your continued patience.
