### 2022-10-31 1674f23
Enhance find_grids() function and add find_dates() function
