### 2022-10-31 ee8e96f
Use find_pairs() function when pairs are not defined
