### 2022-11-02 9a67ecd
Enhance detrending to use lazy compulation
