+++ pygmtsar/pygmtsar/SBAS_detrend.py
@@ -4,7 +4,8 @@ from .SBAS_unwrap import SBAS_unwrap
 
 class SBAS_detrend(SBAS_unwrap):
 
-    def detrend_parallel(self, pairs=None, n_jobs=-1, **kwargs):
+    def detrend_parallel(self, pairs=None, n_jobs=-1, interactive=False, **kwargs):
+        import pandas as pd
         from tqdm.auto import tqdm
         import joblib
 
@@ -16,23 +17,31 @@ class SBAS_detrend(SBAS_unwrap):
         def func(pair, **kwargs):
             #print (f'**kwargs {kwargs}')
             grid = self.open_grids([pair], 'unwrap', interactive=False)
-            
-            kwargs1 = {k:v for k,v in kwargs.items() if k in ['wavelength', 'truncate', 'resolution_meters', 'debug']}
-            #print (f'kwargs1 {kwargs1}')
-            if 'wavelength' in kwargs1:
-                out1 = self.degaussian(grid[0], **kwargs1)
-            else:
-                out1 = grid[0]
-        
+
+            # without any processing options return the input as is
+            out = grid[0]
+
+            if 'wavelength' in kwargs:
+                kwargs1 = {k:v for k,v in kwargs.items() if k in ['wavelength', 'truncate', 'resolution_meters', 'debug']}
+                #print (f'kwargs1 {kwargs1}')
+                out = self.degaussian(out, **kwargs1)
+
             kwargs2 = {k:v for k,v in kwargs.items() if k in ['fit_intercept', 'fit_dem', 'fit_coords',
                                                               'resolution_meters', 'debug']}
             #print (f'kwargs2 {kwargs2}')
-            out2 = self.detrend(out1, **kwargs2)
-        
-            self.save_grids([out2], 'detrend', interactive=False)
+            # ignore detrending when all the fits are disabled
+            out = self.detrend(out, **kwargs2)
 
-        with self.tqdm_joblib(tqdm(desc='Detrending and Saving', total=len(pairs))) as progress_bar:
-            joblib.Parallel(n_jobs=1)(joblib.delayed(func)(pair, **kwargs) for pair in pairs)
+            if interactive:
+                return out
+            self.save_grids([out], 'detrend', interactive=False)
+
+        label = 'Detrending and Saving' if not interactive else 'Detrending'
+        with self.tqdm_joblib(tqdm(desc=label, total=len(pairs))) as progress_bar:
+            results = joblib.Parallel(n_jobs=n_jobs)(joblib.delayed(func)(pair, **kwargs) for pair in pairs)
+        
+        if interactive:
+            return results
 
     def detrend(self, dataarray, fit_intercept=True, fit_dem=True, fit_coords=True,
                 resolution_meters=90, debug=False):
@@ -47,12 +56,14 @@ class SBAS_detrend(SBAS_unwrap):
         from sklearn.linear_model import LinearRegression
         from sklearn.pipeline import make_pipeline
         from sklearn.preprocessing import StandardScaler
-    
+
         def postprocessing(out):
             return out.astype(np.float32).rename('detrend')
 
         # check the simplest case
-        assert fit_intercept or fit_dem or fit_coords, 'All the detrending options disable, function does nothing'
+        if not fit_intercept and not fit_dem and not fit_coords:
+            print ('NOTE: All the detrending options disable, function does nothing')
+            return dataarray
 
         # check simple case
         if fit_intercept and not fit_dem and not fit_coords:
@@ -64,6 +75,8 @@ class SBAS_detrend(SBAS_unwrap):
         decimator = self.pixel_decimator(resolution_meters=resolution_meters, grid=dataarray, debug=debug)
         # decimate
[…]
