### 2022-11-03 #60 @lwasser
Hi, Alexey.

First and foremost I wanted to say thank you for all of the energy that you’ve put into this review and as a maintainer for pyGMTstar. We know that working with the GMT tools is not easy. And we appreciate your expertise as a maintainer and your efforts to make accessing sentinel data easier for scientists.

I want to start by saying we are not rejecting your package. However, in order to review your package, we require some significant changes to its current structure.

A foundational goal for pyOpenSci is [continued maintenance of our packages](https://www.pyopensci.org/peer-review-guide/software-peer-review-guide/author-guide.html#do-you-plan-to-continue-to-maintain-your-package). We are not a journal per se, rather we are an ecosystem that curates lists of vetted, peer reviewed tools. Our partnership with JOSS supports the Journal / publication component if publishing is all that you desire. As such, for every package that comes to pyOpenSci, we need to carefully consider whether we can support future maintenance of the tool in case a maintainer steps down. Mainainters with the best intentions burn out, get new jobs, etc. The issue of long-term maintenance and dependability of packages is a problem that pyOpenSci seeks to address.

Your package, as designed, would be challenging for us to support. Its core structure depends upon two repositories being updated and staying in sync. Syncing files across two repos would be extremely difficult for a new maintainer to take on.

Another foundational goal for pyOpenSci is to facilitate contributions from the community to our packages. Contributions from others who are not working on pyGMTstar would also be incredibly challenging given the current structure of pyGMTStar package.
Here is what we’d need to review pyGMTStar
While we don’t have this specifically documented yet, we’d prefer that the code base from gmtstar be in some way wrapped as a dependency or vendored in some way. Some have suggested a git submodule might work. In this scenario you would push all updates to the gmtstar repo directly that related to the C codebase. pyGMTstar would NOT be a fork.
Related to the above, the README file in the case of pyGMTstar should contain a few sentences at the top information about how your package relates to GMTstar (gmtstar is effectively vendored via your fork). And how the code base between the two repos is synced (or not synced). In the case of this review, had we understood how gmtstar was vendored and synced earlier on, we would have been able to have this discussion with you sooner.
Development instructions: We require that every package contains instructions for contributing. A CONTRIBUTING.md file that contains this information about how you expect contributions to work, and in this case with some indication of dependencies that might need to be installed, should be provided in the README.  [This link onGitHub is one we plan to use in future issue templates as an initial check that you have some of the basic files in your repo that helps a user navigate.](https://github.com/gmtsar/gmtsar/community) Some packages have more robust contributing documentation with issue templates and broader guides on how the issue/pr/ workflow operates in the project; this level of documentation is not expected by us but we see it and appreciate it when it’s there. These are not new standards. These standards are ones that even GitHub embraces. I am working hard to better document things on our end for future submissions.
Repository name: For user clarity we prefer that the repository name is the name of the package. If that is not possible, at the minimum it’s confusing for users to have two repositories named gmtstar - one of which is actually pygmtstar. Please know that this is NOT A NEW rule. However it’s taking me some time to edit our packaging guide.
README: that same PR that I linked to above has more specifics about what we want to see in your readme. I do understand that David missed some elements that existed in your documentation / repository. However, one core element of a good README file is that it is organized in a way that it directs users to all of those elements. This too has always existed in our docs. I am working on a rewrite now to make that more clear.

Please know that right now I am the sole full time employee of pyOpenSci. And I am lucky to be working with volunteers like David and others every day to build this project. This is only my second month here full time. Unfortunately I can only do so much each day which is why some of the things I wrote above aren’t fully in writing - yet. This is the first package we have considered reviewing that has a C library as its base and that is also vendored via a fork.

Again, we think you’ve done incredible work here. We are not rejecting you. Rather we are saying that right now, the way the package is structured conflicts with our goals as an organization. And thus we are unable to review it. However, in the future, if you are able to work on the items above, we could consider it for review.

I understand and respect if you disagree with anything that i’ve written above. But I need to adhere to the goals and mission of the organization as I make decisions. If you have any followup questions, I'm happy to answer them here. If there are no questions, we will close this issue (for now). You are welcome to re-open it in the future or submit a new issue.

You might also consider looking at JOSS in case they do have the technical expertise to review your package as it is now.
For now i've marked the issue out-of-scope for us as it's out of our technical scope to review.

Thank you for understanding what we are trying to achieve at pyOpenSci, Alexey. We appreciate your hard work on this package.
Respectfully, Leah
