### 2022-11-05 5eedc40
Rewrite nearest_grid() function to be lazy and use the distance in meters instead of pixels
