### 2022-11-05 65f5779
Fix unwrapping function to work for bog rasters in interactive mode and allow to use user supplied phase and coherency grids.
