### 2022-11-06 6846818
Add check for input grid radar coordinates
