### 2022-11-06 cad2039
Enhance intf_ra2ll() function to save all the input coordinates
