### 2022-11-07 4e2dd99
Change open_grids() logic to apply the user-defined function for post-processing
