### 2022-11-07 ee073fa
Fix assert message
