### 2022-11-08 2c7e2f8
Increase default searh radius to fill empty grid values
