### 2022-11-08 5e75abd
Change nearest_grid() function to pixel-based distance
