### 2022-11-08 73a2460
Revert to pixel-based nearest_grid() interpolation function
