### 2022-11-08 169a3e4
Update test case S1A_Stack_CPGF_T173 for the recent changes
