+++ .github/workflows/macos.yml
@@ -53,11 +53,6 @@ jobs:
         path: tests/S1A_Stack_CPGF_T173.tar.gz
         key: S1A_Stack_CPGF_T173.tar.gz
         restore-keys: S1A_Stack_CPGF_T173
-    - name: Download dataset
-      working-directory: tests
-      run: |
-        wget -c http://topex.ucsd.edu/gmtsar/tar/S1A_Stack_CPGF_T173.tar.gz
-        tar xvzf S1A_Stack_CPGF_T173.tar.gz -C .
     - name: Run test
       working-directory: tests
       run: |
