### 2022-11-08 c5f19a5
For Ubuntu tests workflow save test outputs as GitHub Artefacts
