### 2022-11-09 8eb849f
Update test case for the recent code changes and add wrapper shell script to download the dataset and start the test
