### 2022-11-09 f2fefd8
Enhance interpolation function to work on the both radar and geographic grids
