### 2022-11-09 45b2f9b
Enhance mask processing for open_grids() function to auto-define when the mask should be applied
