### 2022-11-09 5f3dad9
Add shell script directive to stop execution on any error
