### 2022-11-09 714105a
Update documentation to remove subswath argument for some functions where this argument is excluded to simplify the API
