+++ docs/README.md
@@ -242,15 +242,12 @@ def get_dem(self, subswath=None, geoloc=False, buffer_degrees=0.02):
 #### Manage topography in radar coordinates (combined GMTSAR wrapper and PyGMTSAR original)
 
 ```
-def topo_ra_parallel(self, n_jobs=-1, method='cubic'):
+def topo_ra_parallel(self, interactive=False):
 
 		Build WGS84 DEM in radar coordinates for interferogram processing.
 		
     Args:
-    		n_jobs: number of parallel processing jobs. n_jobs=-1 means all the processor cores used.
-        method: interpolation method applied when the original DEM is lower than defined
-        "resolution_meters". 'cubic' interpolation prefered because it saves well the DEM
-        spatial spectrum.
+    		interactive: boolean flag to return a lazy delayed Xarray dataarray (True) or use chunked computing to build and save the array to file with progress indicator.
 
 		Returns:
         None
@@ -260,9 +257,9 @@ def topo_ra_parallel(self, n_jobs=-1, method='cubic'):
 ```
 
 ```
-def get_topo_ra(self, subswath=None):
+def get_topo_ra(self):
 
-		Return WGS84 DEM in radar coordinates as just one or list of Xarray Dataarray.
+		Return WGS84 DEM in radar coordinates as one or list of Xarray Dataarray depending of the subswaths count.
 		
     Args:
     		subswath - optional subswath number.
@@ -273,9 +270,6 @@ def get_topo_ra(self, subswath=None):
     Examples:
     		Get DEM for all the processed subswaths:
         topo_ra = sbas.get_topo_ra()
-        
-        Get DEM for a single subswath IW1:
-        topo_ra = sbas.get_topo_ra(1)
 ```
 
 #### Framing (combined GMTSAR wrapper and PyGMTSAR original)
@@ -554,37 +548,9 @@ def unwrap_parallel(self, pairs, n_jobs=-1, threshold=None, conf=None, func=None
 #### Detrending (PyGMTSAR original)
 
 ```
-def detrend(self, subswath, pair=None, wavelength=None, topo_ra=None, truncate=3.0, approximate=True, fit_intercept=True, fit_dem=True, fit_coords=True, debug=False):
-
-		Detrend and get output for a single unwrapped interferogram combining topography and linear components removal plus Gaussian filtering.
-		
-    Args:
-    		subswath: optional argument to define a subswath.
-    		pairs: mandatory list of dates pairs (baseline pairs).
-    		n_jobs: number of parallel processing jobs. n_jobs=-1 means all the processor cores used.
-    		wavelength: optional cut-off wavelength for Gaussian filter in meters.
-    		topo_ra: optional user-defined Xarray Dataarray topography grid instead of main one.
-    		truncate: optional Gaussian filter window in sigmas.
-    		approximate: optional boolean flag to use approximate Gaussian filter calculation for large gammas.
-    		fit_intercept: optional boolean flag to drop mean value (plane).
-    		fit_dem: optional boolean flag to detrend topography.
-    		fit_coords: optional boolean flag to detrend linear coordinate components.
-    		debug: boolean flag to print debug information.
-
-		Returns:
-        2D Xarray Dataarray
-
-    Examples:
-    		Simplest detrending:
-    		unwrap_detrended = sbas.detrend(pair.values[0] if isinstance(pairs, pd.DataFrame) else pair[0])
-    		Detrend ionospreric effects and solid Earth's tides on large area:
-    		unwrap_detrended = sbas.detrend(pairs.values[0] wavelength=12000)
-```
-
-```
[…]
