+++ .github/workflows/pypi.yml
@@ -57,6 +57,7 @@ jobs:
       run: |
         # upgrade package to resolve dependencies
         pip3 install click --upgrade
+        pip3 install matplotlib seaborn
         pip3 install pygmtsar
     - name: Cache dataset
       uses: actions/cache@v3
@@ -69,6 +70,12 @@ jobs:
       run: |
         export PATH=$PATH:/home/runner/work/gmtsar/gmtsar/bin
         sh S1A_Stack_CPGF_T173.sh
+    - name: Archive test results
+      uses: actions/upload-artifact@v3
+      with:
+        name: plots
+        path: tests/*.jpg
+        if-no-files-found: error
 
   S1A_2016_Kumamoto_Earthquake:
 
@@ -116,47 +123,19 @@ jobs:
       run: |
         # upgrade package to resolve dependencies
         pip3 install click --upgrade
+        pip3 install matplotlib seaborn
         pip3 install pygmtsar
-    - name: Download dataset
-      working-directory: tests
-      run: |
-        # download iCloud datasets
-        cat > icloud_urls.txt <<- END
-        https://www.icloud.com/iclouddrive/04946wsi9N5g0W6beTl_KUvQg#S1A_IW_SLC__1SSV_20160408T091355_20160408T091430_010728_01001F_83EB.zip
-        https://www.icloud.com/iclouddrive/059EHU1iazg_CDEBWd7CPjXnQ#S1A_IW_SLC__1SSV_20160420T091355_20160420T091423_010903_010569_F9CE.zip
-        https://www.icloud.com/iclouddrive/042d2NP0tDVFz_N39jCu5yyVA#S1A_Kumamoto_Earthquake_ESA_Sentinel_1_Toolbox.jpg
-        END
-        cut -f5 -d '/' icloud_urls.txt | cut -f1 -d '#' | sed -E 's/(.*)/{"shortGUIDs":[{"value":"\1"}]}/' > requests.txt
-        cut -f5 -d '/' icloud_urls.txt | cut -f2 -d '#' > fnames.txt
-        paste -d ' ' requests.txt fnames.txt  | while read LINE
-        do
-            request=$(echo "$LINE"| cut -d ' ' -f1)
-            fname=$(echo "$LINE"| cut -d ' ' -f2)
-            echo "request: $request fname: $fname"
-            url=$(curl -s 'https://ckdatabasews.icloud.com/database/1/com.apple.cloudkit/production/public/records/resolve' \
-            --data-raw "${request}" --compressed | jq -r '.results[0].rootRecord.fields.fileContent.value.downloadURL')
-            curl -so "${fname}" "${url}"
-        done
-        # unpack to data directory
-        mkdir -p data
-        unzip -j -n S1A_IW_SLC__1SSV_20160408T091355_20160408T091430_010728_01001F_83EB.zip \
-            '*.SAFE/*/s1?-iw1-slc-vv-*-001.xml' -d data
-        unzip -j -n S1A_IW_SLC__1SSV_20160408T091355_20160408T091430_010728_01001F_83EB.zip \
-            '*.SAFE/*/s1?-iw1-slc-vv-*-001.tiff' -d data
-        unzip -j -n S1A_IW_SLC__1SSV_20160420T091355_20160420T091423_010903_010569_F9CE.zip \
-            '*.SAFE/*/s1?-iw1-slc-vv-*-001.xml' -d data
-        unzip -j -n S1A_IW_SLC__1SSV_20160420T091355_20160420T091423_010903_010569_F9CE.zip \
-            '*.SAFE/*/s1?-iw1-slc-vv-*-001.tiff' -d data
-        # cleanup
-        rm *.zip
-        # check
-        df -h
-        ls data
[…]
