### 2022-11-09 11eb4ca
Enhance test script to produce more plots
