+++ tests/S1A_Stack_CPGF_T173.py
@@ -6,8 +6,11 @@
 from pygmtsar import SBAS
 import numpy as np
 import xarray as xr
+import pandas as pd
 import os
 from dask.distributed import Client
+import matplotlib.pyplot as plt
+import matplotlib
 
 #Define Parameters
 BACKUPDIR    = 'backup'
@@ -77,6 +80,7 @@ if __name__ == '__main__':
     print (sbas.open_grids(pairs, 'detrend'))
     # PyGMTSAR SBAS Displacement
     sbas.sbas_parallel(pairs)
+
     # output
     def interp(da):
         return sbas.nearest_grid(da).where(~np.isnan(validmask))
@@ -131,18 +135,42 @@ if __name__ == '__main__':
         assert os.path.exists(os.path.join(DATADIR,     orbit)), 'Original file not found'
         assert not os.path.exists(os.path.join(WORKDIR, orbit)), 'Excessive file found'
 
-    # plot
-    corrs = sbas.open_grids(pairs, 'corr')
-    fg = corrs.plot.imshow(
-        col="pair",
-        col_wrap=3, size=4, aspect=1.2,
-        clim=(0, 0.8), cmap='gray'
-    )
-    fg.set_axis_labels('Range', 'Azimuth')
-    fg.set_ticks(max_xticks=5, max_yticks=5, fontsize='medium')
-    fg.fig.suptitle('Correlation', y=1.05, fontsize=24)
-    fg.fig.savefig('Correlation.jpg', dpi=150, pil_kwargs={"quality": 95})
-    
+    # plots
+    dem = sbas.get_dem()
+    plt.figure(figsize=(12,4), dpi=300)
+    dem.plot.imshow(cmap='gray', vmin=0)
+    sbas.to_dataframe().plot(color='orange', alpha=0.2, ax=plt.gca())
+    plt.title('Sentinel1 Frame on DEM', fontsize=18)
+    plt.savefig('Sentinel1 Frame on DEM.jpg', dpi=300, pil_kwargs={"quality": 95})
+
+    plt.figure(figsize=(12,4), dpi=300)
+    ax = plt.gca()
+    lines = [[(row.ref_timeline,row.ref_baseline),(row.rep_timeline,row.rep_baseline)]
+             for row in baseline_pairs.itertuples()]
+    lc = matplotlib.collections.LineCollection(lines, colors='#30a2da', linewidths=0.5)
+    ax.add_collection(lc)
+    ax.autoscale()
+    ax.margins(0.5)
+    bs1 = baseline_pairs[['ref_timeline','ref_baseline','ref_date']].values
+    bs2 = baseline_pairs[['rep_timeline','rep_baseline','rep_date']].values
+    df = pd.DataFrame(np.concatenate([bs1, bs2]), columns=['timeline','baseline','date']).drop_duplicates()
+    plt.scatter(x=df.timeline, y=df.baseline)
+    for x,y,label in df.values:
+        plt.annotate(label, (x,y), textcoords="offset points", xytext=(35,3), ha='center',
+                     c='red' if label == MASTER else 'black')
+    ax.set_xlabel('Timeline', fontsize=16)
+    ax.set_ylabel('Perpendicular Baseline, [m]', fontsize=16)
+    ax.set_title('SBAS Baseline', fontsize=18)
+    plt.savefig('SBAS Baseline.jpg', dpi=300, pil_kwargs={"quality": 95})
+
+    topo_ra = sbas.get_topo_ra()
[…]
