### 2022-11-09 d8a70df
Update Kumamoto test case and add plots
