### 2022-11-09 6e80962
Allow to apply mask two times as it needed for some cases
