### 2022-11-09 cd98188
Add more plots for Crete test case
