### 2022-11-10 16b3dad
Remove annotations from progressbar code to work everywhere
