### 2022-11-10 ecadaf0
Upgrade setup-python@v3 to setup-python@v4 in CI tests
