+++ .github/workflows/pypi.yml
@@ -73,7 +73,7 @@ jobs:
     - name: Archive test results
       uses: actions/upload-artifact@v3
       with:
-        name: plots
+        name: Plots (${{ matrix.os }}, ${{ matrix.python-version }})
         path: tests/*.jpg
         if-no-files-found: error
 
@@ -133,7 +133,7 @@ jobs:
     - name: Archive test results
       uses: actions/upload-artifact@v3
       with:
-        name: plots
+        name: Plots (${{ matrix.os }}, ${{ matrix.python-version }})
         path: tests/*.jpg
         if-no-files-found: error
 
@@ -193,6 +193,6 @@ jobs:
     - name: Archive test results
       uses: actions/upload-artifact@v3
       with:
-        name: plots
+        name: Plots (${{ matrix.os }}, ${{ matrix.python-version }})
         path: tests/*.jpg
         if-no-files-found: error
