### 2022-11-12 4deb972
Increase PyPI package version
