### 2022-11-12 4ddc317
Try to fix too long downloading progress incicator prints for CI tests
