### 2022-11-12 8a5141c
Remove Python 3.7 and use 3.8, 3.9, 3.10, 3.11 for PyPI tests
