### 2022-11-12 848f97f
Use Python 3.8, 3.9, 3.10 only for PyPI CI tests. Note: dask distributed does not work with Python 3.7 and is not ready for Python 3.11
