### 2022-11-12 bb7f2b7
Add workaround for Python 3.7 (Google Colab)
