### 2022-11-12 1d2d45a
Add case name to CI test plots
