### 2022-11-12 5593732
Try to revert Python 3.7 CI test runs
