### 2022-11-12 a544e7a
Try to fix weird error on Python 3.8 and 3.9 (while it works on 3.10) for Kumamoto CI test only
