### 2022-11-12 08edaa2
Run CI tests with force cleanup flag
