### 2022-11-12 194d0b0
Fix CI test issues related to high RAM consumption
