### 2022-11-13 03d5c85
Update test scripts
