### 2022-11-15 ed258ed
Enhance sat_look_parallel() function
