### 2022-11-18 494b9db
Enhance rounding for pixel size in meters
