### 2022-11-18 0703bb6
Enhance SBAS baseline pairs analysis to do it before images aligning
