### 2022-11-23 74c0843
Enhance large data grids processing in containers
