+++ pygmtsar/pygmtsar/SBAS_topo_ra.py
@@ -15,7 +15,7 @@ class SBAS_topo_ra(SBAS_trans):
 
         # trans.dat - file generated by llt_grid2rat (r a topo lon lat)"
         trans_dat = self.get_trans_dat(subswath)
-    
+
         trans_blocks_extents = self.get_trans_dat_blocks_extents(subswath, n_jobs=n_jobs)
 
         # define topo_ra grid
@@ -33,7 +33,7 @@ class SBAS_topo_ra(SBAS_trans):
         def calc_topo_ra(azi, rng):
             # check thr arguments
             assert azi.shape == rng.shape, f'ERROR: {azi.shape} != {rng.shape}'
-        
+
             # check the selected area bounds
             ymin, ymax, xmin, xmax = azi.min(), azi.max(), rng.min(), rng.max()
             #print ('ymin, ymax', ymin, ymax, 'xmin, xmax', xmin, xmax)
@@ -48,9 +48,9 @@ class SBAS_topo_ra(SBAS_trans):
             blocks_eles = []
             for iy, ix in blocks[:,:2].astype(int):
                 #print ('iy, ix', iy, ix)
-                block_azi = trans_dat.azi.data.blocks[iy, ix]
-                block_rng = trans_dat.rng.data.blocks[iy, ix]
-                block_ele = trans_dat.ele.data.blocks[iy, ix]
+                block_azi = trans_dat.azi.data.blocks[iy, ix].persist()
+                block_rng = trans_dat.rng.data.blocks[iy, ix].persist()
+                block_ele = trans_dat.ele.data.blocks[iy, ix].persist()
                 #print ('block_ele', block_ele.shape)
 
                 blocks_azis.append(block_azi.reshape(-1))
@@ -59,16 +59,16 @@ class SBAS_topo_ra(SBAS_trans):
             blocks_azis = np.concatenate(blocks_azis)
             blocks_rngs = np.concatenate(blocks_rngs)
             blocks_eles = np.concatenate(blocks_eles)
-        
+
             # build index tree - dask arrays computes automatically
             source_yxs = np.stack([blocks_azis, blocks_rngs], axis=1)
             tree = cKDTree(source_yxs, compact_nodes=False, balanced_tree=False)
-        
+
             # query the index tree
             target_yxs = np.stack([azi.reshape(-1), rng.reshape(-1)], axis=1)
             d, inds = tree.query(target_yxs, k = 1, workers=1)
             # compute dask array to prevent ineffective index loockup on it
-            matrix = blocks_eles.compute()[inds].reshape(azi.shape)
+            matrix = np.asarray(blocks_eles)[inds].reshape(azi.shape)
             return matrix
 
         # xarray wrapper
