### 2022-11-24 632b27e
Enhance CI test scripts
