### 2022-11-25 4d76f82
Add check for NetCDF incorrect size due to issues in h5netcdf library
