+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -22,3 +22,5 @@ If applicable, add screenshots to help explain your problem.
 
 **Processing Log**
 If applicable, attach a log file from terminal output
+
+[Please note that if no new responses are made over six months from users, the issue will get closed. If the report involves a simple fix in the code, the change could be done quickly; if it involves developing new features, do expect them to be done over the long run. Under both cases, the report will remain open until the issue is addressed one way or another.]
