+++ .github/ISSUE_TEMPLATE/general-help.md
@@ -24,3 +24,7 @@ Attach a log file output from your terminal.
 
 **Screenshot**
 If applicable, attach a Screenshot to help us understand your case
+
+[Please note that if no new responses are made over six months from users, the issue will get closed]
+
+[Also if this is a general question that's not particlarly related to GMTSAR, such as "how to correct atmospheric noise" or "how to intepret my interferogram", consider using the discussions page]
